% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.GeoTimeseries}
\alias{aggregate.GeoTimeseries}
\title{Aggregate the Metrics of a GeoTimeseries.}
\usage{
\method{aggregate}{GeoTimeseries}(x, by = kGeo, FUN = base::sum,
  metrics = NULL, ...)
}
\arguments{
\item{x}{a GeoTimeseries object.}

\item{by}{(character vector) name(s) of column(s) by which to group.}

\item{FUN}{(function) function to apply to each metric column.}

\item{metrics}{(character vector) metrics to aggregate. Default is all
metrics.}

\item{...}{optional arguments passed to FUN.}
}
\value{
A data.frame object.
}
\description{
Aggregate the Metrics of a GeoTimeseries.
}
\note{
Uses \code{aggregate.data.frame} to do the aggregation. This function
omits rows that have missing values in the '\code{by}' columns.
}
\seealso{
AggregateTimeseries.
}

