% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatetimeseries.R
\name{AggregateTimeseries.GeoTimeseries}
\alias{AggregateTimeseries.GeoTimeseries}
\title{Convert a GeoTimeseries to a weekly or monthly GeoTimeseries. (Note:
intended to transform daily data.)}
\usage{
\method{AggregateTimeseries}{GeoTimeseries}(obj, freq, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{freq}{desired frequency: 'weekly' or 'monthly'.}

\item{...}{ignored.}
}
\value{
A GeoTimeseries object with the weekly or monthly metrics
  aggregated. The sums are associated with the last day of the week
  (which is by our definition _Sunday_), or the last day of the month.
}
\description{
Convert a GeoTimeseries to a weekly or monthly GeoTimeseries. (Note:
intended to transform daily data.)
}
\note{
'Weekly' frequency: each day in the time series is mapped to the next
Sunday, then aggregated. 'Monthly' frequency: each day in the time
series is mapped to the last day of the month, then aggregated. No
check is made about the current frequency. This works best on daily
data. So it is possible to attempt to change a 'monthly' frequency back
to 'weekly', but this simply re-maps the last day of the month to the
next Sunday.
}
\seealso{
AggregateTimeseries (generic), aggregate.GeoTimeseries.
}

