% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatetimeseries.R
\name{AggregateTimeseries}
\alias{AggregateTimeseries}
\alias{AggregateTimeseries.GeoTimeseries}
\title{Aggregate the metrics of a time series object over specified time
intervals.}
\usage{
AggregateTimeseries(obj, freq = c("weekly", "monthly"), ...)

\method{AggregateTimeseries}{GeoTimeseries}(obj, freq, ...)
}
\arguments{
\item{obj}{an object.}

\item{freq}{(string) 'weekly' or 'monthly' aggregation.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of the same class as 'obj'.
}
\description{
Aggregate the metrics of a time series object over specified time
intervals.
}
\note{
\itemize{
  \item{'Weekly' frequency}: each day in the time series is mapped
    to the next Sunday, then aggregated.
  \item{'Monthly' frequency}: each day in the time series is mapped to
    the last day of the month, then aggregated. No check is made about
    the current frequency. This works best on daily data. So it is
    possible to attempt to change a 'monthly' frequency back to 'weekly',
    but this simply re-maps the last day of the month to the next Sunday.
}
}
\seealso{
\code{\link{aggregate.GeoTimeseries}}.
}

