% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbrroasanalysisdata.R
\name{as.GBRROASAnalysisData}
\alias{as.GBRROASAnalysisData}
\title{Coerces the object to a GBRROASAnalysisData object.}
\usage{
as.GBRROASAnalysisData(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A GBRROASAnalysisData object.
}
\description{
Coerces the object to a GBRROASAnalysisData object.
}
\note{
A generic S3 method.
}
\seealso{
as.GBRROASAnalysisData.GeoExperimentData.
}

