% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_geoexperimentdata.R
\name{as.GeoExperimentData.GeoTimeseries}
\alias{as.GeoExperimentData.GeoTimeseries}
\title{Coerces a GeoTimeseries object to a GeoExperiment object.}
\usage{
\method{as.GeoExperimentData}{GeoTimeseries}(obj, strict = TRUE, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object with the additional columns 'period',
'geo.group', and 'assignment'. If any of these columns are missing,
the corresponding columns in the resulting object will be 'NA'.}

\item{strict}{(flag) if FALSE, the additional columns are optional and no
error is thrown if any of them is missing;}

\item{...}{ignored.}
}
\value{
A GeoExperimentData object.
}
\description{
Coerces a GeoTimeseries object to a GeoExperiment object.
}
\seealso{
as.GeoExperimentData (generic).
}

