% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_geotimeseries.R
\name{as.GeoTimeseries}
\alias{as.GeoTimeseries}
\title{Coerce an object to a GeoTimeseries object.}
\usage{
as.GeoTimeseries(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A GeoTimeseries object.
}
\description{
Coerce an object to a GeoTimeseries object.
}

