% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as.matrix.GeoTimeseries}
\alias{as.matrix.GeoTimeseries}
\title{Coerces a response metric of a GeoTimeseries to a matrix representation
with geos in rows, dates in columns.}
\usage{
\method{as.matrix}{GeoTimeseries}(x, response, ...)
}
\arguments{
\item{x}{a GeoTimeseries object.}

\item{response}{(string) name of the response metric to use.}

\item{...}{ignored.}
}
\value{
A numeric matrix of the response metric, geos in rows, dates in
  columns.
}
\description{
Coerces a response metric of a GeoTimeseries to a matrix representation
with geos in rows, dates in columns.
}

