% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as.matrix.TBRQuantiles}
\alias{as.matrix.TBRQuantiles}
\title{Extracts the real-valued matrix of quantiles from a TBRQuantiles object.}
\usage{
\method{as.matrix}{TBRQuantiles}(x, ...)
}
\arguments{
\item{x}{a TBRQuantiles object.}

\item{...}{ignored.}
}
\value{
A real-valued matrix of the quantiles. The column names are the
  same as those in the TBRQuantiles object.
}
\description{
Extracts the real-valued matrix of quantiles from a TBRQuantiles object.
}

