% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tbranalysisdata.R
\name{as.TBRAnalysisData}
\alias{as.TBRAnalysisData}
\title{Coerces the object to a TBRAnalysisData object.}
\usage{
as.TBRAnalysisData(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A TBRAnalysisData object.
}
\description{
Coerces the object to a TBRAnalysisData object.
}
\note{
A generic S3 method.
}
\seealso{
as.TBRAnalysisData.GeoExperimentData.
}

