% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tbrplotdata.R
\name{as.TBRPlotData}
\alias{as.TBRPlotData}
\title{Coerces an object to a TBRPlotData object.}
\usage{
as.TBRPlotData(obj, panels, periods, quantiles = c(0.1, 0.9), ...)
}
\arguments{
\item{obj}{an object to coerce.}

\item{panels}{(vector of strings or NULL) names of the panels to be plotted.
By default, all available panels.}

\item{periods}{(vector of strings or NULL) names of the periods to show. By
default, all default periods.}

\item{quantiles}{(real vector of length 2) lower and upper quantiles of the
credible interval to show.}

\item{...}{further arguments passed to the methods.}
}
\value{
A TBRPlotData object.
}
\description{
Coerces an object to a TBRPlotData object.
}

