% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tbrplotdata.R
\name{as.TBRPlotData.TBRAnalysisFitTbr1}
\alias{as.TBRPlotData.TBRAnalysisFitTbr1}
\title{Coerces a TBRAnalysisFitTbr1 object to a TBRPlotData object.}
\usage{
\method{as.TBRPlotData}{TBRAnalysisFitTbr1}(obj,
  panels = GetTBRPlotPanelNames(), periods = c("pretest", "prediction"),
  quantiles = c(0.1, 0.9), ...)
}
\arguments{
\item{obj}{an object to coerce.}

\item{panels}{(vector of strings) names of the panels to be plotted.}

\item{periods}{(vector of strings) names of the periods to show.}

\item{quantiles}{(real vector of length 2) lower and upper quantiles of the
credible interval to show.}

\item{...}{ignored.}
}
\value{
A TBRPlotData object.
}
\description{
Coerces a TBRAnalysisFitTbr1 object to a TBRPlotData object.
}

