% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckForBadValues}
\alias{CheckForBadValues}
\title{Check for bad values in given columns.}
\usage{
CheckForBadValues(x, columns, CHECK, good = TRUE, what = "invalid", ...)
}
\arguments{
\item{x}{(data frame) any data frame.}

\item{columns}{(character) columns to check.}

\item{CHECK}{(function) function that returns a logical vector (of length
\code{nrow(x)}).}

\item{good}{(logical) the value or values that indicate a 'good' value.
Other values returned by 'CHECK' will be 'bad' and throw an error.}

\item{what}{(string) a string to display in the informative message.}

\item{...}{further arguments passed on to function 'CHECK'.}
}
\value{
If all tests pass, returns NULL invisibly. Otherwise, an
  'assertError' is thrown.
}
\description{
Check for bad values in given columns.
}

