% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckForDuplicateRows}
\alias{CheckForDuplicateRows}
\title{Tests whether the values in rows of given columns are not duplicated.}
\usage{
CheckForDuplicateRows(x, columns)
}
\arguments{
\item{x}{a data frame.}

\item{columns}{(character) vector.}
}
\value{
If all tests pass, returns TRUE invisibly. Otherwise, an
  'assertError' is thrown.
}
\description{
Tests whether the values in rows of given columns are not duplicated.
}

