% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckForMapping}
\alias{CheckForMapping}
\title{Tests whether the mapping is consistent within a data frame.}
\usage{
CheckForMapping(x, from, to)
}
\arguments{
\item{x}{a data frame.}

\item{from}{(character) vector of column names.}

\item{to}{(character) vector of column names. Typically these are different
from those in 'from'.}
}
\value{
If all tests pass, returns TRUE invisibly. Otherwise, an
  'assertError' is thrown.
}
\description{
Tests whether the mapping is consistent within a data frame.
}
\note{
Checks that no rows in 'x[from]' map to two different values of
'x[to]'.
}

