% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckForTypes}
\alias{CheckForTypes}
\title{Checks whether the vector values in the given list satisfies the specified
types.}
\usage{
CheckForTypes(x, checklist)
}
\arguments{
\item{x}{a named list or a data frame.}

\item{checklist}{(named list) mapping from names to functions that check
whether the functions are correct or not.}
}
\value{
invisible NULL.
}
\description{
Checks whether the vector values in the given list satisfies the specified
types.
}

