% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckGeoGroupNumber}
\alias{CheckGeoGroupNumber}
\title{Check that the specified geo group number is present in the data.}
\usage{
CheckGeoGroupNumber(geo.group, values)
}
\arguments{
\item{geo.group}{(an integer) geo group number (>= 1). (NA throws an error).}

\item{values}{(vector) values to check against.}
}
\value{
TRUE, invisibly; as a side effect may throw an assertion error.
}
\description{
Check that the specified geo group number is present in the data.
}
\note{
For checking arguments within a function. Outputs the name of the
variable in error messages.
}

