% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckPeriodNumbers}
\alias{CheckPeriodNumbers}
\title{Check that the specified period numbers are present in the data.}
\usage{
CheckPeriodNumbers(period, values)
}
\arguments{
\item{period}{(integer vector): one or more period numbers. (NA throws an
error).}

\item{values}{(vector) values to check against.}
}
\value{
TRUE invisibly if the checks pass. As a side effect will terminate
  with an error if the checks fail.
}
\description{
Check that the specified period numbers are present in the data.
}
\note{
For checking arguments within a function. Outputs the name of the
variable in error messages.
}

