% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbrroasanalysisfit_gbr1.R
\name{coef.GBRROASAnalysisFitGbr1}
\alias{coef.GBRROASAnalysisFitGbr1}
\title{Returns the model coefficients.}
\usage{
\method{coef}{GBRROASAnalysisFitGbr1}(object, ...)
}
\arguments{
\item{object}{a GBRROASAnalysisFitGbr1 object.}

\item{...}{ignored.}
}
\value{
A named numeric vector with the model coefficients from the GBR
  analysis.
}
\description{
Returns the model coefficients.
}

