% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{ConcatItems}
\alias{ConcatItems}
\title{Concatenate items of a vector, optionally quoting each.}
\usage{
ConcatItems(x, quote = "'", collapse = ", ", max.output = Inf)
}
\arguments{
\item{x}{(atomic vector) an atomic vector, coerced to character.}

\item{quote}{(string) a quote character to use.}

\item{collapse}{(string) string to use for collapsing the components of 'x'.}

\item{max.output}{(integer) maximum number of items of 'x' to output.}
}
\value{
A string.
}
\description{
Concatenate items of a vector, optionally quoting each.
}

