% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostrata.R
\name{CountRandomizations}
\alias{CountRandomizations}
\title{Counts the total numbers of randomizations in a GeoStrata object.}
\usage{
CountRandomizations(geostrata, show.warnings = TRUE, log.scale = FALSE)
}
\arguments{
\item{geostrata}{a GeoStrata object.}

\item{show.warnings}{(flag) if \code{TRUE}, shows a warning when a stratum
is not compatible with the group.ratios.}

\item{log.scale}{(flag) if \code{TRUE}, returns the result on the log.scale.}
}
\value{
An integer vector of the same length as the number of strata in
geostrata (excluding stratum 0, if it exists). The i-th coordinate
corresponds to the total number of possible randomizations for the i-th
stratum.
}
\description{
Counts the total numbers of randomizations in a GeoStrata object.
}
\note{
A warning is issued if one or more of the stratas are not compatible
with the group.ratios.
}

