% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateposterior.R
\name{cumsum.PosteriorSimulations}
\alias{cumsum.PosteriorSimulations}
\title{Returns simulations of the posterior cumulative joint distribution.}
\usage{
\method{cumsum}{PosteriorSimulations}(x)
}
\arguments{
\item{x}{a PosteriorSimulations object.}
}
\value{
A PosteriorSimulations object.
}
\description{
Returns simulations of the posterior cumulative joint distribution.
}

