% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentassignment.R
\name{DefaultTreatmentAssignment}
\alias{DefaultTreatmentAssignment}
\title{Generates the default treatment assignment condition.}
\usage{
DefaultTreatmentAssignment()
}
\value{
A TreatmentAssignment object.
}
\description{
Generates the default treatment assignment condition.
}
\note{
Period 1 of group 2 is assigned a change (some intervention).
}

