% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogbrroasanalysis.R
\name{DoGBRROASAnalysis.GBRROASAnalysisData}
\alias{DoGBRROASAnalysis.GBRROASAnalysisData}
\title{Performs a ROAS Analysis on a GBRROASAnalysisData object.}
\usage{
\method{DoGBRROASAnalysis}{GBRROASAnalysisData}(obj, ...)
}
\arguments{
\item{obj}{a GBRROASAnalysisData object.}

\item{...}{ignored.}
}
\value{
A GBRROASAnalysisFit object.
}
\description{
Performs a ROAS Analysis on a GBRROASAnalysisData object.
}
\seealso{
DoGBRROASAnalysis (generic), DoGBRROASAnalysis.GeoExperimentData.
}

