% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doroasanalysis.R
\name{DoROASAnalysis}
\alias{DoROASAnalysis}
\title{Performs a ROAS analysis using the given model.}
\usage{
DoROASAnalysis(obj, model, ...)
}
\arguments{
\item{obj}{an object.}

\item{model}{(string) id of the model to use. Available models are 'gbr1',
'tbr1'.}

\item{...}{further arguments passed to or from other models.}
}
\value{
A GBRROASAnalysisFit (for 'gbr1') or a TBRROASAnalysisFit object
  (for 'tbr1').
}
\description{
Performs a ROAS analysis using the given model.
}
\note{
Dispatches the right method for the given model type.
}
\seealso{
\code{\link{DoGBRROASAnalysis}}, \code{\link{DoTBRROASAnalysis}}.
}

