% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doroaspreanalysis.R
\name{DoROASPreanalysis.GeoExperimentPreanalysisData}
\alias{DoROASPreanalysis.GeoExperimentPreanalysisData}
\title{Predicts the standard deviation of the posterior of the iROAS estimator by
simulating experiments using historical data.}
\usage{
\method{DoROASPreanalysis}{GeoExperimentPreanalysisData}(obj, response, prop.to,
  n.sims = 1000, models = GetModelIds(), FUN = base::lapply, ...)
}
\arguments{
\item{obj}{a GeoExperimentPreanalysisData object.}

\item{response}{(string) name of the metric column to use as the response
variable.}

\item{prop.to}{(string) an existing name of the column in proportion to
which the spend change will be distributed across the geos.}

\item{n.sims}{(integer or NA) number of simulations to do. Note: if 'geos'
is a GeoAssignment object, NA is allowed, and the actual number of
simulations will be the number of all possible pseudo-data sets; if
n.sims is given, only the first n.sims data sets will be simulated
(useful for testing only).}

\item{models}{(vector of nonempty strings) one or more analysis model ids,
to apply to each simulated data set.}

\item{FUN}{(function) a lapply-compatible function to use for generating the
simulations. Use this for a parallel version of lapply to speed up
the process. Note: the function only needs to be able to iterate
along a vector of integers.}

\item{...}{arguments passed to FUN.}
}
\value{
A ROASPreanalysisFit object.
}
\description{
Predicts the standard deviation of the posterior of the iROAS estimator by
simulating experiments using historical data.
}
\note{
Simulates experiments without resorting to regenerating time series.
Segments of the original time series are used to create
'pseudo-timeseries' of the length of the experiment. These simulated
geo experiment data sets are produced by calls to
'SimulateGeoExperimentData'.

The object that is returned contains the simulated raw numbers. Use
the 'summary' method for a more user-friendly output.
}

