% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doroaspreanalysis.R
\name{DoROASPreanalysis.GeoTimeseries}
\alias{DoROASPreanalysis.GeoTimeseries}
\title{Predicts the standard deviation of the posterior of the iROAS estimator by
simulating experiments using historical data.}
\usage{
\method{DoROASPreanalysis}{GeoTimeseries}(obj, period.lengths, geos,
  recycle = TRUE, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{period.lengths}{a vector of length 3, denoting the lengths (in days)
of pre-period, test, and the cooldown periods, respectively. The test
period must be at least 7 days; the pre-period must be at least as
long as the test period; the cooldown period can be 0 or more days.}

\item{geos}{(GeoStrata or GeoAssignment object) object to use for choosing
the geo groups; if it is a GeoAssignment object, the geo assignment
will be fixed. If it is a GeoStrata object, method 'Randomize' will
be used to obtain a geo assignment.}

\item{recycle}{(flag) if TRUE, uses an augmented data set in order to create
a larger time series, reusing data from the head of the time series.}

\item{...}{arguments passed to
DoROASPreanalysis.GeoExperimentPreanalysisData.}
}
\value{
A ROASPreanalysisFit object.
}
\description{
Predicts the standard deviation of the posterior of the iROAS estimator by
simulating experiments using historical data.
}
\note{
This is a simple wrapper, combining calls to
'GeoExperimentPreanalysisData' and 'DoROASPreanalysis'.
}

