% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbranalysis.R
\name{DoTBRAnalysis.GeoExperimentData}
\alias{DoTBRAnalysis.GeoExperimentData}
\title{Performs a TBR Causal Effect Analysis on a GeoExperimentData object.}
\usage{
\method{DoTBRAnalysis}{GeoExperimentData}(obj, model, ...)
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{model}{(string) model to use. See the generic method for possible
values.}

\item{...}{arguments passed on to the method performing the analysis.}
}
\value{
A TBRAnalysisFit object.
}
\description{
Performs a TBR Causal Effect Analysis on a GeoExperimentData object.
}
\note{
Calls 'as.TBRAnalysisData' and passes the 'TBRAnalysisData' object to
the analysis method.
}
\seealso{
DoTBRAnalysis (generic), DoTBRAnalysis.TBRAnalysisData.
}

