% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbranalysis.R
\name{DoTBRAnalysis}
\alias{DoTBRAnalysis}
\alias{DoTBRAnalysis.GeoExperimentData}
\alias{DoTBRAnalysis.TBRAnalysisData}
\title{Executes a TBR Causal Effect Analysis.}
\usage{
DoTBRAnalysis(obj, model, ...)

\method{DoTBRAnalysis}{GeoExperimentData}(obj, model, ...)

\method{DoTBRAnalysis}{TBRAnalysisData}(obj, model, ...)
}
\arguments{
\item{obj}{an object.}

\item{model}{(string) model to use.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A TBRAnalysisFit object.
}
\description{
Executes a TBR Causal Effect Analysis.
}
\note{
Dispatcher of methods.
The actual method to execute the 'tbr1' method is \code{DoTBRAnalysisTbr1}.
}
\seealso{
\code{\link{DoTBRAnalysisTbr1}}, \code{\link{DoROASAnalysis}},
\code{\link{DoGBRROASAnalysis}}.
}

