% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbranalysis.R
\name{DoTBRAnalysis.TBRAnalysisData}
\alias{DoTBRAnalysis.TBRAnalysisData}
\title{Performs a TBR Analysis on a TBRAnalysisData object.}
\usage{
\method{DoTBRAnalysis}{TBRAnalysisData}(obj, model, ...)
}
\arguments{
\item{obj}{a TBRAnalysisData object.}

\item{model}{(string) model to use. See the generic method for possible
values.}

\item{...}{Arguments passed on to the method performing the analysis.}
}
\value{
A TBRAnalysisFit object.
}
\description{
Performs a TBR Analysis on a TBRAnalysisData object.
}
\seealso{
DoTBRAnalysis (generic), DoTBRAnalysis.GeoExperimentData.
}

