% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbranalysis_tbr1.R
\name{DoTBRAnalysisTbr1}
\alias{DoTBRAnalysisTbr1}
\title{Performs a TBR Analysis using the 2-parameter linear model 'tbr1'.}
\usage{
DoTBRAnalysisTbr1(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A TBRAnalysisFit object.
}
\description{
Performs a TBR Analysis using the 2-parameter linear model 'tbr1'.
}
\note{
A generic S3 method. Users should use 'DoTBRAnalysis' instead,
specifying the 'model' parameter.
}
\seealso{
DoTBRAnalysis.TBRAnalysisData, DoTBRAnalysisTbr1.TBRAnalysisData
}

