% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbranalysis_tbr1.R
\name{DoTBRAnalysisTbr1.TBRAnalysisData}
\alias{DoTBRAnalysisTbr1.TBRAnalysisData}
\title{Performs a TBR Analysis using the 2-parameter linear model 'tbr1'.}
\usage{
\method{DoTBRAnalysisTbr1}{TBRAnalysisData}(obj, ...)
}
\arguments{
\item{obj}{a TBRAnalysisData object.}

\item{...}{ignored.}
}
\value{
A TBRAnalysisFit object.
}
\description{
Performs a TBR Analysis using the 2-parameter linear model 'tbr1'.
}
\note{
Users should use 'DoTBRAnalysis' instead.
}
\seealso{
DoTBRAnalysis.TBRAnalysisData, DoTBRAnalysisTbr1.TBRAnalysisData
}

