% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbrroasanalysis.R
\name{DoTBRROASAnalysis.GeoExperimentData}
\alias{DoTBRROASAnalysis.GeoExperimentData}
\title{Performs a TBR ROAS Analysis.}
\usage{
\method{DoTBRROASAnalysis}{GeoExperimentData}(obj, model, response, cost,
  n.sims = 1e+05, ...)
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{model}{(string) id of the TBR to use.}

\item{response}{(string) name of the column of the response variable.}

\item{cost}{(string) name of the column of the cost variable, or
alternatively a real number specifying the exact total incremental
cost (spend change).}

\item{n.sims}{(integer) number of simulations to draw. If cost is constant
and n.sims == 1, only the point estimates will be calculated. If cost
is not constant and n.sims == 1, an error is thrown.}

\item{...}{arguments passed to function 'DoTBRAnalysis'.}
}
\value{
A TBRROASAnalysisFit object.
}
\description{
Performs a TBR ROAS Analysis.
}
\seealso{
DoTBRROASAnalysis.GeoExperimentData.
}

