% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbrroasanalysis.R
\name{DoTBRROASAnalysis}
\alias{DoTBRROASAnalysis}
\alias{DoTBRROASAnalysis.GeoExperimentData}
\title{Performs a TBR ROAS Analysis.}
\usage{
DoTBRROASAnalysis(obj, model, response, cost, n.sims = 1e+05, ...)

\method{DoTBRROASAnalysis}{GeoExperimentData}(obj, model, response, cost,
  n.sims = 1e+05, ...)
}
\arguments{
\item{obj}{an object.}

\item{model}{(string) id of the TBR to use.}

\item{response}{(string) name of the column of the response variable.}

\item{cost}{(string) name of the column of the cost variable, or
alternatively a real number specifying the exact incremental cost
(spend change).}

\item{n.sims}{(integer) number of simulations to draw.}

\item{...}{arguments passed to function 'DoTBRAnalysis'.}
}
\value{
A TBRROASAnalysisFit object.
}
\description{
Performs a TBR ROAS Analysis.
}
\seealso{
DoTBRAnalysis, DoGBRROASAnalysis
}

