% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateincremental.R
\name{EstimateIncremental}
\alias{EstimateIncremental}
\title{Estimates the difference of a metric between the test and counterfactual
between the pre and the post periods.}
\usage{
EstimateIncremental(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object representing the adjusted incremental values
}
\description{
Estimates the difference of a metric between the test and counterfactual
between the pre and the post periods.
}
\note{
A generic S3 method.
}
\seealso{
EstimateIncremental.GBRROASAnalysisData.
}

