% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractexperimentperiods.R
\name{ExtractExperimentPeriods.GeoTimeseries}
\alias{ExtractExperimentPeriods.GeoTimeseries}
\title{Extracts a ExperimentPeriods object from a GeoTimeseries.}
\usage{
\method{ExtractExperimentPeriods}{GeoTimeseries}(obj, strict = TRUE, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object with the column 'period'.}

\item{strict}{(flag) if FALSE, the function returns NULL if the column
'period' does not exist. Otherwise, throws an error.}

\item{...}{ignored.}
}
\value{
An ExperimentPeriods object.
}
\description{
Extracts a ExperimentPeriods object from a GeoTimeseries.
}
\seealso{
ExtractExperimentPeriods (generic).
}

