% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeoassignment.R
\name{ExtractGeoAssignment.GeoExperimentData}
\alias{ExtractGeoAssignment.GeoExperimentData}
\title{Extracts the associated GeoAssignment object.}
\usage{
\method{ExtractGeoAssignment}{GeoExperimentData}(obj, ...)
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{...}{ignored.}
}
\value{
The GeoAssignment object associated with the GeoExperimentData
  object, if it exists; if it has not been set, returns NULL.
}
\description{
Extracts the associated GeoAssignment object.
}
\seealso{
ExtractGeoAssignment.GeoTimeseries.
}

