% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeoassignment.R
\name{ExtractGeoAssignment.GeoTimeseries}
\alias{ExtractGeoAssignment.GeoTimeseries}
\title{Attempts to extract a GeoAssignment object from a GeoTimeseries.}
\usage{
\method{ExtractGeoAssignment}{GeoTimeseries}(obj, strict = TRUE, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object with the column 'geo.group'.}

\item{strict}{(flag) if FALSE, the function returns NULL if the column
'geo.group' does not exist. Otherwise, throws an error.}

\item{...}{ignored.}
}
\value{
A GeoAssignment object.
}
\description{
Attempts to extract a GeoAssignment object from a GeoTimeseries.
}
\note{
Finds all unique pairs of ('geo', 'geo.group') in the GeoTimeseries.
'geo.group' can have missing values.
}
\seealso{
ExtractGeoAssignment (generic).
}

