% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeoassignment.R
\name{ExtractGeoAssignment}
\alias{ExtractGeoAssignment}
\title{Extracts a GeoAssignment object.}
\usage{
ExtractGeoAssignment(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed on to methods.}
}
\value{
A GeoAssignment object.
}
\description{
Extracts a GeoAssignment object.
}
\note{
A generic S3 method.
}
\seealso{
ExtractGeoAssignment.GeoTimeseries.
}

