% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeostrata.R
\name{ExtractGeoStrata.GeoExperimentData}
\alias{ExtractGeoStrata.GeoExperimentData}
\title{Extract a GeoStrata object from a GeoExperimentData object.}
\usage{
\method{ExtractGeoStrata}{GeoExperimentData}(obj, volume = NULL, ...)
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{volume}{(string) name of a metric in the GeoExperimentData over which
to generate the 'volume' column.}

\item{...}{arguments passed to 'GeoStrata'.}
}
\value{
A GeoStrata object, obtained by first extracting a Geos object from
  'obj'.
}
\description{
Extract a GeoStrata object from a GeoExperimentData object.
}

