% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeostrata.R
\name{ExtractGeoStrata.GeoTimeseries}
\alias{ExtractGeoStrata.GeoTimeseries}
\title{Extract a GeoStrata object from a GeoTimeseries.}
\usage{
\method{ExtractGeoStrata}{GeoTimeseries}(obj, volume = NULL, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{volume}{(string) name of a metric in the GeoTimeseries over which to
generate the 'volume' column.}

\item{...}{arguments passed to 'GeoStrata'.}
}
\value{
A GeoStrata object, obtained by first extracting a Geos object from
  'obj'.
}
\description{
Extract a GeoStrata object from a GeoTimeseries.
}

