% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeostrata.R
\name{ExtractGeoStrata}
\alias{ExtractGeoStrata}
\title{Extract a GeoStrata object from an object.}
\usage{
ExtractGeoStrata(obj, ...)
}
\arguments{
\item{obj}{some object for which the method 'ExtractGeos' exists.}

\item{...}{other arguments passed to the methods.}
}
\value{
A GeoStrata object.
}
\description{
Extract a GeoStrata object from an object.
}

