% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracttreatmentassignment.R
\name{ExtractTreatmentAssignment.GeoTimeseries}
\alias{ExtractTreatmentAssignment.GeoTimeseries}
\title{Extracts a TreatmentAssignment object from a GeoTimeseries object.}
\usage{
\method{ExtractTreatmentAssignment}{GeoTimeseries}(obj, strict = TRUE, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object with the columns 'period', 'geo.group',
and 'assignment'.}

\item{strict}{(flag) if FALSE, the function returns NULL if either of the
columns 'geo.group' or 'period' does not exist. Otherwise, throws an
error.}

\item{...}{ignored.}
}
\value{
A TreatmentAssignment object.
}
\description{
Extracts a TreatmentAssignment object from a GeoTimeseries object.
}
\note{
A well-defined (period, group) pair (i.e., neither of them is
missing) in the data frame implies that the corresponding (date, geo)
pair is part of the experiment and *must* be therefore associated a
treatment condition. Otherwise, if a date or a geo is not part of the
experiment, the (date, geo) pair *must* have *no* treatment condition
assignment. In other words, any (period, group) pair that maps to a
non-missing treatment assignment must have no missing values.
Conversely, any (period, group) pair with a missing value must map to a
missing treatment assignment.
}
\seealso{
ExtractTreatmentAssignment (generic).
}

