% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracttreatmentassignment.R
\name{ExtractTreatmentAssignment}
\alias{ExtractTreatmentAssignment}
\alias{ExtractTreatmentAssignment.GeoTimeseries}
\title{Extracts a TreatmentAssignment object.}
\usage{
ExtractTreatmentAssignment(obj, ...)

\method{ExtractTreatmentAssignment}{GeoTimeseries}(obj, strict = TRUE, ...)
}
\arguments{
\item{obj}{An object.}

\item{...}{further arguments passed to or from other methods.}

\item{strict}{(flag) if FALSE, the function returns NULL if either of the
columns 'geo.group' or 'period' does not exist. Otherwise, throws an
error.}
}
\value{
A TreatmentAssignment object.
}
\description{
Extracts a TreatmentAssignment object.
}
\note{
\code{ExtractTreatmentAssignment.GeoTimeseries}: \code{obj} is
supposed to have the columns 'period', 'geo.group', and 'assignment'.\cr

A well-defined (period, group) pair (i.e., neither of them is missing) in
the data frame implies that the corresponding (date, geo) pair is part of
the experiment and \emph{must} be therefore associated a treatment
condition. Otherwise, if a date or a geo is not part of the experiment, the
(date, geo) pair \emph{must} have \emph{no} treatment condition
assignment. In other words, any (period, group) pair that maps to a
non-missing treatment assignment must have no missing values.  Conversely,
any (period, group) pair with a missing value must map to a missing
treatment assignment.
}

