% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{FormatText}
\alias{FormatText}
\title{Compose text strings conditional on values of an object.}
\usage{
FormatText(x, ..., quote = "'")
}
\arguments{
\item{x}{an integer-valued numeric scalar or a logical vector. In the former
case, indicates a number of items; in the latter case, indicates
which items in a vector were affected. May have the 'names' attribute
set.}

\item{...}{one or more character vectors that are pasted together
(collapsed) after concatenating the vectors one after another.}

\item{quote}{the quote character to use.}
}
\value{
A character string.
}
\description{
Compose text strings conditional on values of an object.
}
\note{
The function replaces the following special character patterns:
\itemize{
\item{\code{\{a|b}\}}: 'a' if sum(x) == 1 and 'b' otherwise.
\item{\code{\{z|a|b\}}}: 'z' if sum(x) == 0, otherwise works like
{a|b}.
\item{\code{$N}}: sum(x).
\item{\code{$P}}: sprintf("%.1f", 100 * mean(x)).
\item{\code{$L}}: length(x). } The following patterns are replaced with
comma-separated lists of:
\itemize{
  \item{\code{$w}}: which(x).
  \item{\code{$W}}: which(x), each item quoted.
  \item{\code{$x}}: names(x)[which(x)].
  \item{\code{$X}}: names(x)[which(x)], each item quoted.
}
Exception: These four patterns output only up to the Kth item, where K =
getOption('FormatTextMaxOutput', default=7L).
}
\examples{
\dontrun{
FormatText(n, "There {is|are} {no|one|$N} item{|s}.")
FormatText(is.na(x), "Found $N NAs ($P\% of all $L) in rows $w")}
}

