% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbrroasanalysisdata.R
\name{GBRROASAnalysisData}
\alias{GBRROASAnalysisData}
\title{Constructs a GBRROASAnalysisData object.}
\usage{
GBRROASAnalysisData(x)
}
\arguments{
\item{x}{a _plain_ 'data.frame' or an object that has to be coercible to a
_plain_ 'data.frame' with the required columns 'geo', which is a
character vector; numeric columns 'resp.pre', 'resp.test',
'cost.pre', 'cost.test', and a logical (TRUE/FALSE) column 'control'.
Optionally the data frame can have other columns, whose names can not
start with a dot ('.'). No missing values are allowed.}
}
\value{
A GBRROASAnalysisData object, which is similarly a 'data.frame'
  with the same columns as described above.
}
\description{
Constructs a GBRROASAnalysisData object.
}
\note{
'GBRROASAnalysisData' is a 'data.frame', with the required columns,
\itemize{
\item\code{geo} a 'character'-valued vector of Geo IDs.
\item\code{resp.pre} numeric values of the response metric in the
pre-test period.
\item\code{resp.test} numeric values of the response metric in the test
period.
\item\code{cost.pre} numeric values of the cost metric in the pre-test
period.
\item\code{cost.test} numeric values of the response metric in the test
period.
\item\code{control} : indicator ('TRUE'/'FALSE') of which columns are
in the Control group.} In addition, optionally any other user-definable
columns. The column 'geo' is the primary key; it is guaranteed that no
rows with duplicate geos exist. The object includes the following
fields stored in the attribute 'info':
\itemize{
  \item\code{keys}: names of the primary key columns ('geo').
    \item\code{required}: names of the required columns.
    \item\code{metrics}: names of the columns representing metric data.
}
}

