% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostrata.R
\name{.GenerateStrata}
\alias{.GenerateStrata}
\title{[internal] Generate stratum numbers along a given vector.}
\usage{
.GenerateStrata(geo.group, group.ratios)
}
\arguments{
\item{geo.group}{(integer vector of length equal to number of geos) a vector
of geo group numbers (may be NAs).}

\item{group.ratios}{(integer vector of length equal to number of geo groups)
vector of ratios of the sizes of each group.}
}
\value{
An integer vector of the same length as 'geo.group', with the
corresponding stratum numbers. '0' signifies a geo that is excluded from the
set. Otherwise the numbers identify the strata. There are no NAs.
}
\description{
[internal] Generate stratum numbers along a given vector.
}
\note{
For internal use; no checking of arguments is done.
}

