% Copyright (C) 2017 Google, Inc.

\name{geoassignment}
\docType{data}
\alias{geoassignment}
\title{Sample Geo Assignment Data}
\usage{data(geoassignment)}
\description{
  A data frame, with columns.
  \describe{
    \item{geo:}{Geo ID.}
    \item{geo.group:}{Geo Group ID.}
  }
}
\source{
  Simulated data.
}
