% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoassignment.R
\name{GeoAssignment}
\alias{GeoAssignment}
\title{Constructs a GeoAssignment object.}
\usage{
GeoAssignment(x)
}
\arguments{
\item{x}{a data frame with columns 'geo' (character) and 'geo.group'
(integer-valued, positive). 'geo.group' indicates the group (1, 2,
...) the corresponding geo belongs to. Other columns are allowed but
not checked. If 'geo' is an integer-valued numeric or factor, it is
coerced to character.}
}
\value{
An object of class 'GeoAssignment'.
}
\description{
Constructs a GeoAssignment object.
}
\note{
The group numbers must be integers (integer-valued numeric are
allowed as input), starting with 1. The column 'geo.group' *can* have
missing values (NA), unlike 'geo', which must have no missing values,
and no duplicates. A missing value in geo.group simply indicates that
the mapping from a geo is not available. When associating the geo
assignment with a GeoExperimentData object, the effect will be the same
as if the geo with 'geo.group'==NA was not in the mapping in the first
place.
}

