% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoexperimentdata.R
\name{GeoExperimentData}
\alias{GeoExperimentData}
\title{Constructs a GeoExperimentData object.}
\usage{
GeoExperimentData(geo.timeseries, periods, geo.assignment, treat.assignment)
}
\arguments{
\item{geo.timeseries}{a GeoTimeseries object.}

\item{periods}{an ExperimentPeriods object, specifying the start dates of
each period in the experiment. Or, 'NULL' if the dates are yet
unknown. Or, leave unspecified to extract this information from the
column 'periods' of the geo.timeseries object; in this case the
column *must* exist in the data frame.}

\item{geo.assignment}{a GeoAssignment object, specifying the mapping from a
geo to a geo group. Or, 'NULL' if the geo assignment is yet unknown.
Or, leave unspecified to extract this information from the column
'geo.group' of the geo.timeseries object; in this case the column
*must* exist in the data frame.}

\item{treat.assignment}{a TreatmentAssignment object, specifying the mapping
from (period, geo) to a treatment assignment condition (treatment
intervention type). Or, 'NULL' if the treatment assignment is yet
unknown. Or, leave unspecified to extract this information from the
column 'assignment' of the geo.timeseries object; in this case the
column *must* exist in the data frame.}
}
\value{
A GeoExperimentData object.
}
\description{
Constructs a GeoExperimentData object.
}
\note{
If any of the arguments 'periods', 'geo.assignment', or
'treat.assignment' is not specified, a column 'period', 'geo.group', or
'assignment', respectively, *must* be present in the 'geo.timeseries'
object. If any of the arguments 'periods', 'geo.assignment', or
'treat.assignment' is NULL, the corresponding column ('period',
'geo.group', or 'assignment') in the resulting GeoExperimentData data
frame will have only 'NA's. If the GeoTimeseries object has columns
'period', 'geo.group', or 'assignment', they will be overwritten by the
values given by the objects 'periods', 'geo.assignment', and
'treat.assignment'. The resulting object *may* have undefined
'periods', 'geo.assignment', or 'treat.assignment'. The missing parts
may be filled in by using the functions 'SetExperimentPeriods',
'SetGeoAssignment', and 'SetTreatmentAssignment'.
}

