% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoxpreanalysisdata.R
\name{GeoExperimentPreanalysisData}
\alias{GeoExperimentPreanalysisData}
\title{Creates a GeoExperimentPreanalysisData object.}
\usage{
GeoExperimentPreanalysisData(obj, period.lengths, geos, recycle = TRUE)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{period.lengths}{an integer-valued vector of length 2 or 3, denoting
the lengths (in days) of pre-period, test, and the (optional)
cooldown periods, respectively. The test period must be at least 7
days; the pre-period must be at least as long as the test period. The
cooldown period can be 0 or more days; if it is zero, it will be
ignored as if there were only 2 periods.}

\item{geos}{(GeoStrata or GeoAssignment object) object to use for choosing
the geo groups at the time of simulating (using
SimulateGeoExperimentData); if 'geos' is a GeoAssignment object, then
the geo assignment will be fixed; iteration; if 'geos' is a
GeoStrata, then the geo assignment will be generated by a call to
'Randomize'.}

\item{recycle}{(flag) if TRUE, creates an augmented data set in order to
create a larger time series, reusing data from the head of the time
series.}
}
\value{
A GeoExperimentPreanalysis object, which inherits from
  GeoExperimentData, with NAs in 'geo.group', 'period', and
  'assignment'. The default treatment assignment (spend change only for
  period 1 and group 2) is stored into the object.
}
\description{
Creates a GeoExperimentPreanalysisData object.
}
\note{
A GeoExperimentPreanalysisData object stores a historical data set
and the information about the length of the experiment periods. It is
used as the generator of pseudo-data sets for simulation purposes. See
also: 'SimulateGeoExperimentData' to generate a GeoExperimentData
object.
}

