% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos.R
\name{Geos}
\alias{Geos}
\title{Constructor for Geos objects.}
\usage{
Geos(x, volume = NULL)
}
\arguments{
\item{x}{a data frame with column 'geo' and optionally other columns. Each
'geo' must be unique.}

\item{volume}{name of the column that represents the 'volume' of the geo.
The proportion of volume is computed. If omitted, the assumed volumes
will be 1 for each geo. The volumes have to be non-negative and sum
up to a positive value.}
}
\value{
An object of class 'Geos', which is a data frame with the column
  'geo', the column 'proportion' and 'volume', and other optional
  columns. The rows are sorted by the descending order of 'volume'.
}
\description{
Constructor for Geos objects.
}

