% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostrata.R
\name{GeoStrata}
\alias{GeoStrata}
\title{Constructor for \code{GeoStrata} objects.}
\usage{
GeoStrata(geos, n.groups = 2, group.ratios = rep(1, length.out = n.groups))
}
\arguments{
\item{geos}{a \code{Geos} object.}

\item{n.groups}{number of groups. At least 2 and at most the number of geos.}

\item{group.ratios}{(integer vector of length n.groups) vector of ratios of
the sizes of each group. By default each group is assumed to have equal
ratios. The sum of these numbers also imply the size of a stratum. For
example, \code{c(2, 1)} implies that group 1 should be 2 times larger than
group 2, and the stratum size is 2 + 1 = 3. Note: the ratios do not have to
be normalized to have greatest common divisor 1. For example, c(4, 2)
implies that the ratio of group sizes is 2:1 but the stratum size is 4 + 2 =
6.}
}
\value{
A \code{GeoStrata} object that inherits from \code{Geos}. There is
an extra column \code{stratum} that indicates the stratum number to be used
in randomization, and column \code{geo.group} for fixing the geo-to-group
mapping.
}
\description{
Constructor for \code{GeoStrata} objects.
}
\details{
\code{GeoStrata} objects are used for (stratified) randomization of
geos into groups. The geos are sorted by their 'volume' (definable by the
user) and then divided into strata of size n.groups (column 'stratum').
This object has also a column \code{geo.group}, which offers the possibility to
fix certain geos to certain groups. By default, this column is filled with
\code{NA}s, indicating that none of the geos are mapped to any groups. The
randomization itself is done by the method \code{Randomize}.\cr

Any individual geo -> geo.group mappings should be fixed by using the
\code{SetGeoGroup<-} method on a \code{GeoStrata} object.\cr

A stratum number \code{0} indicates a geo that is excluded from the scheme
stratification. Any geo that is mapped to group \code{0} will have stratum
number \code{0}; for example if geo \code{2} was omitted (geo.groups were
\code{NA, 0, NA, NA, NA, ...}) with group.ratios \code{c(2, 1)}, the strata
would be assigned as \code{1, 0, 1, 1, 2, 2, 2, 3, 3, 3, 4, ...}\cr

Setting \code{group.ratios} to some other value than the default
\code{1,1,...} enables creating groups that have different sizes. The
stratum size is then determined by the sum of the number in
\code{group.ratios}. For example, \code{group.ratios=c(1, 2)} implies a
ratio of 1:2. Each stratum has size 3; the 3 geos in this stratum are
assigned a random sample with replacement from the set
\code{1,2,2}. Similarly, \code{c(3, 1)} implies that group 1 will be on
average 3 times as large as group 2.\cr
}
\note{
The ratios do not have to be normalized to have greatest common
divisor 1. For example, \code{c(4, 2)} implies that the ratio of group sizes is 2:1
but the stratum size is 4 + 2 = 6.
}
\seealso{
\code{\link{Randomize}}, \code{\link{SetGeoGroup<-}}.
}

