% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeogroup.R
\name{GetGeoGroup.GeoAssignment}
\alias{GetGeoGroup.GeoAssignment}
\title{Returns the geo group of a geo in a GeoStrata object.}
\usage{
\method{GetGeoGroup}{GeoAssignment}(obj, geo = NULL, ...)
}
\arguments{
\item{obj}{a GeoAssignment object.}

\item{geo}{(character vector or NULL) names of geos for which to obtain the
geo group ids.}

\item{...}{ignored.}
}
\value{
A named integer-valued vector, with the geo names in the names
  attribute, mapping geos to geo group ids.
}
\description{
Returns the geo group of a geo in a GeoStrata object.
}
\note{
If a specified 'geo' does not exist in the GeoAssignment object, the
corresponding geo group will be an NA.
}

