% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeogroup.R
\name{GetGeoGroup.GeoExperimentData}
\alias{GetGeoGroup.GeoExperimentData}
\title{Returns the geo group of a geo in a GeoExperimentData object.}
\usage{
\method{GetGeoGroup}{GeoExperimentData}(obj, geo = NULL, ...)
}
\arguments{
\item{obj}{a GeoExperiment object.}

\item{geo}{(character vector or NULL) names of geos for which to obtain the
geo group ids.}

\item{...}{ignored.}
}
\value{
A named integer-valued vector, with the geo names in the names
  attribute, mapping geos to geo group ids.
}
\description{
Returns the geo group of a geo in a GeoExperimentData object.
}

