% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeogroup.R
\name{GetGeoGroup}
\alias{GetGeoGroup}
\alias{GetGeoGroup.GeoAssignment}
\alias{GetGeoGroup.GeoExperimentData}
\alias{GetGeoGroup.GeoStrata}
\title{Returns the geo-to-geo group mapping.}
\usage{
GetGeoGroup(obj, geo = NULL, ...)

\method{GetGeoGroup}{GeoAssignment}(obj, geo = NULL, ...)

\method{GetGeoGroup}{GeoExperimentData}(obj, geo = NULL, ...)

\method{GetGeoGroup}{GeoStrata}(obj, geo = NULL, ...)
}
\arguments{
\item{obj}{an object with the columns 'geo' and 'geo.group'.}

\item{geo}{(character vector or NULL) names of geos for which to obtain the
geo group ids.}

\item{...}{other arguments passed on to the methods.}
}
\value{
A named integer-valued vector, with the geo names in the names
  attribute, mapping geos to geo group ids.
}
\description{
Returns the geo-to-geo group mapping.
}
\note{
If a specified 'geo' does not exist in the GeoAssignment object, the
corresponding geo group will be an NA.
}

