% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeonames.R
\name{GetGeoNames.GeoAssignment}
\alias{GetGeoNames.GeoAssignment}
\title{Extracts the names of the geos from a GeoAssignment object.}
\usage{
\method{GetGeoNames}{GeoAssignment}(obj, groups = NULL)
}
\arguments{
\item{obj}{a GeoAssignment object.}

\item{groups}{(NULL or integer vector) id number(s) of the groups whose geos
to obtain, or NULL for all geos. NA is allowed.}
}
\value{
A character vector of unique geo identifiers, sorted by their name.
  May be empty if none of the groups specified are found in the object.
}
\description{
Extracts the names of the geos from a GeoAssignment object.
}

