% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeonames.R
\name{GetGeoNames.GeoExperimentData}
\alias{GetGeoNames.GeoExperimentData}
\title{Extracts the unique names of the geos from a GeoAssignment object.}
\usage{
\method{GetGeoNames}{GeoExperimentData}(obj, groups = NULL)
}
\arguments{
\item{obj}{a GeoTimeseries (including GeoExperimentData) object.}

\item{groups}{(NULL or integer vector) id number(s) of the groups whose geos
to obtain, or NULL for all geos. NA is allowed.}
}
\value{
A character vector of unique geo identifiers, sorted. Can be empty
  if there are no geos matching 'groups'. Will be empty if the geo
  assignment object does not exist in the object.
}
\description{
Extracts the unique names of the geos from a GeoAssignment object.
}

