% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeonames.R
\name{GetGeoNames.Geos}
\alias{GetGeoNames.Geos}
\title{Extracts the unique names of the geos from a Geos object.}
\usage{
\method{GetGeoNames}{Geos}(obj, groups = NULL)
}
\arguments{
\item{obj}{a Geos object.}

\item{groups}{should be NULL, otherwise an error occurs.}
}
\value{
A character vector of unique geo identifiers, sorted. Cannot be
  empty.
}
\description{
Extracts the unique names of the geos from a Geos object.
}

