% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeonames.R
\name{GetGeoNames.GeoStrata}
\alias{GetGeoNames.GeoStrata}
\title{Extracts the unique names of the geos from a GeoStrata object.}
\usage{
\method{GetGeoNames}{GeoStrata}(obj, groups = NULL)
}
\arguments{
\item{obj}{a GeoStrata object.}

\item{groups}{(NULL, integer vector) id number(s) of the groups whose geos
to obtain, or NULL for all geos. NA is allowed.}
}
\value{
A character vector of unique geo identifiers, sorted.
}
\description{
Extracts the unique names of the geos from a GeoStrata object.
}

