% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeonames.R
\name{GetGeoNames.GeoTimeseries}
\alias{GetGeoNames.GeoTimeseries}
\title{Extracts the unique names of the geos from a GeoAssignment object.}
\usage{
\method{GetGeoNames}{GeoTimeseries}(obj, groups = NULL)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{groups}{should be NULL, otherwise an error occurs.}
}
\value{
A character vector of unique geo identifiers, sorted. Can be empty
  if there are no geos matching 'groups'.
}
\description{
Extracts the unique names of the geos from a GeoAssignment object.
}

