% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{GetInfo}
\alias{GetInfo}
\title{Get a value of the 'info' attribute of an object.}
\usage{
GetInfo(obj, field)
}
\arguments{
\item{obj}{any R object.}

\item{field}{(string) name of the field the value of which to retrieve.}
}
\value{
The requested value (object).
}
\description{
Get a value of the 'info' attribute of an object.
}

