% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messaging.R
\name{GetMessageContextString}
\alias{GetMessageContextString}
\title{Get the current message context string.}
\usage{
GetMessageContextString()
}
\value{
The current message context string. If not available, returns an
  empty string ("").
}
\description{
Get the current message context string.
}
\note{
Pastes the strings together with '>'.
}

