% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{GetModelIds}
\alias{GetModelIds}
\title{Returns the list of available model identifiers.}
\usage{
GetModelIds()
}
\value{
A character vector of the available model IDs. The 'names'
  attribute shows the methodology ID (gbr, tbr).
}
\description{
Returns the list of available model identifiers.
}

