% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateroasanalysis.R
\name{.GetPreanalysisSummary}
\alias{.GetPreanalysisSummary}
\alias{.GetPreanalysisSummary.GBRROASAnalysisFitGbr1}
\alias{.GetPreanalysisSummary.TBRROASAnalysisFit}
\title{[internal] Derives a short summary of the quantities for the preanalysis.}
\usage{
.GetPreanalysisSummary(obj, i, geo.assignment, sim)

\method{.GetPreanalysisSummary}{TBRROASAnalysisFit}(obj, i, geo.assignment, sim)

\method{.GetPreanalysisSummary}{GBRROASAnalysisFitGbr1}(obj, i, geo.assignment,
  sim)
}
\arguments{
\item{obj}{an object.}

\item{i}{(integer) index of the data set.}

\item{geo.assignment}{(a \code{GeoAssignment} object) the actual
geo assignment of the data set.}

\item{sim}{(integer >= 0) number of the simulation.}
}
\value{
A \code{ROASAnalysisFit} object.
}
\description{
[internal] Derives a short summary of the quantities for the preanalysis.
}
\note{
No integrity checking of the arguments is done.
}

