% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tbrplotdata.R
\name{.GetTBRDataFrameForGgplot}
\alias{.GetTBRDataFrameForGgplot}
\title{[internal] Arranges a TBRPlotData or TBRROASPlotData object into a data
frame for plotting.}
\usage{
.GetTBRDataFrameForGgplot(obj, panels, periods, lower, upper, panel.info)
}
\arguments{
\item{obj}{a TBRPlotData or TBRROASPlotData object.}

\item{panels}{names of the panels to plot. Available names are those in
\code{(kTBRPlotPanels)}.}

\item{periods}{(vector of strings) names of the periods to show.}

\item{lower}{(0 < number < upper) lower quantile.}

\item{upper}{(lower < number < 1) upper quantile.}

\item{panel.info}{(list) information about the data frame columns for each
of the available panels.}
}
\value{
A data frame with the columns:
  \itemize{
    \item\code{date} date.
    \item\code{observed} an observed time series.
    \item\code{predicted} a predicted time series.
    \item\code{lower} lower bound of the predicted time series.
      \item\code{upper} lower bound of the predicted time series.
      \item\code{panel.label} panel label to be shown in the plot.
}
}
\description{
[internal] Arranges a TBRPlotData or TBRROASPlotData object into a data
frame for plotting.
}

