% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{GetTBRPlotPanelNames}
\alias{GetTBRPlotPanelNames}
\title{Returns the names of available TBR panels in the default order of
plotting.}
\usage{
GetTBRPlotPanelNames()
}
\value{
A character vector of the available panel IDs.
}
\description{
Returns the names of available TBR panels in the default order of
plotting.
}

