% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweeklyaverages.R
\name{GetWeeklyAverages.GeoTimeseries}
\alias{GetWeeklyAverages.GeoTimeseries}
\title{Calculates the weekly averages of all metrics per geo.}
\usage{
\method{GetWeeklyAverages}{GeoTimeseries}(obj, na.rm = TRUE, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{na.rm}{(flag) remove NAs?}

\item{...}{ignored.}
}
\value{
A data frame with one row per 'geo', and the weekly averages of
  each metric in the columns. Sorted by the name of the geo.
}
\description{
Calculates the weekly averages of all metrics per geo.
}
\note{
Does not handle incomplete weeks. Each week is supposed to have the
complete total sales of the week. If for example the first and last
weeks of a daily data set are incomplete, the weekly average will be
underestimated. Works for both weekly and daily data. The averages are
calculated by first calculating the weekly sums for each geo, and then
taking the averages over the weeks. NAs are removed by default.
}

