% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweeklyaverages.R
\name{GetWeeklyAverages}
\alias{GetWeeklyAverages}
\title{Calculates the weekly averages of the metrics per geo.}
\usage{
GetWeeklyAverages(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{arguments passed on to the methods.}
}
\value{
A data frame with one row per 'geo', and the weekly averages of
  each metric in the columns.
}
\description{
Calculates the weekly averages of the metrics per geo.
}
\note{
A generic S3 method.
}
\seealso{
GetWeeklyAverages.GeoTimeseries.
}

