% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tbrplotdata.R
\name{.GetWeeklyShadedBackgroundDataFrameForGgplot}
\alias{.GetWeeklyShadedBackgroundDataFrameForGgplot}
\title{[internal] Create a data.frame to plot a weekly shaded background to
highlight weekly patterns.}
\usage{
.GetWeeklyShadedBackgroundDataFrameForGgplot(plot.data)
}
\arguments{
\item{plot.data}{the output of .GetTBRDataFrameForGgplot.}
}
\value{
A data.frame with the columns:
  \itemize{
    \item\code{xmin}.
    \item\code{panel.label}.
    \item\code{ymin}.
    \item\code{ymax}.
    \item\code{xmax}.
}
}
\description{
[internal] Create a data.frame to plot a weekly shaded background to
highlight weekly patterns.
}

