% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.character.vector}
\alias{is.character.vector}
\title{Tests whether 'x' is a nonempty character vector.}
\usage{
is.character.vector(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a character vector of positive length,
  otherwise FALSE.
}
\description{
Tests whether 'x' is a nonempty character vector.
}

