% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.empty.string}
\alias{is.empty.string}
\title{Tests each component of 'x' to see whether the string is empty.}
\usage{
is.empty.string(x)
}
\arguments{
\item{x}{(character) a vector to test.}
}
\value{
A logical vector of the same length as 'x', with 'TRUE' if and only
  if the string is empty, NA if the string is NA, otherwise FALSE. If
  'x' is not character, throws an error.
}
\description{
Tests each component of 'x' to see whether the string is empty.
}
\note{
NA maps to NA.
}

