% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.inf}
\alias{is.inf}
\title{Tests whether 'x' is Inf or -Inf.}
\usage{
is.inf(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is Inf or -Inf.
}
\description{
Tests whether 'x' is Inf or -Inf.
}

