% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.integer.valued}
\alias{is.integer.valued}
\title{Tests whether 'x' is an integer-valued numeric vector.}
\usage{
is.integer.valued(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is either of type integer, or numeric and
  all components coercible to integer, otherwise FALSE.
}
\description{
Tests whether 'x' is an integer-valued numeric vector.
}
\note{
NAs are allowed.
}

