% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.logical.vector}
\alias{is.logical.vector}
\title{Tests whether 'x' is a nonempty logical vector.}
\usage{
is.logical.vector(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a logical vector of positive length,
  otherwise FALSE.
}
\description{
Tests whether 'x' is a nonempty logical vector.
}

