% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.named.list}
\alias{is.named.list}
\title{Tests whether 'x' is a named list.}
\usage{
is.named.list(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a plain list with a names attribute with
  non-empty strings for names, otherwise FALSE.
}
\description{
Tests whether 'x' is a named list.
}

