% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.nonempty.string}
\alias{is.nonempty.string}
\title{Tests whether 'x' is a nonempty string (and not NA).}
\usage{
is.nonempty.string(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a character vector of length 1 with a
  component that is nonempty and non-NA, otherwise FALSE.
}
\description{
Tests whether 'x' is a nonempty string (and not NA).
}

