% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_vectors.R
\name{is.period.number}
\alias{is.period.number}
\title{Tests each component of the vector 'x' for valid experiment period number.}
\usage{
is.period.number(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
A logical vector of the same length as 'x', with 'TRUE' if and only
  if the component is an integer-valued numeric value, NA if the value
  is NA, otherwise FALSE. If 'x' is not integer-valued, throws an
  error.
}
\description{
Tests each component of the vector 'x' for valid experiment period number.
}
\note{
NA in a component returns NA. These are supposed to be tested
separately. Negative period numbers are allowed.
}

