% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.plain.list}
\alias{is.plain.list}
\title{Tests whether 'x' is a plain list (not an object).}
\usage{
is.plain.list(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a list but not an object. otherwise
  FALSE.
}
\description{
Tests whether 'x' is a plain list (not an object).
}

