% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.real.number}
\alias{is.real.number}
\title{Tests whether 'x' is a real number.}
\usage{
is.real.number(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a numeric scalar, not an NA, and not
  infinite, otherwise FALSE.
}
\description{
Tests whether 'x' is a real number.
}

