% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_vectors.R
\name{is.treatment.assignment}
\alias{is.treatment.assignment}
\title{Tests each component of the vector 'x' for valid treatment assignment
symbols.}
\usage{
is.treatment.assignment(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
A logical vector of the same length as 'x', with 'TRUE' if and only
  if the string is a valid treatment assignment symbol, NA if the
  string is NA, otherwise FALSE. If 'x' is not character, throws an
  error.
}
\description{
Tests each component of the vector 'x' for valid treatment assignment
symbols.
}
\note{
NA in a component returns NA. These are supposed to be tested
separately.
}

