% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{is.vector.of.nonempty.strings}
\alias{is.vector.of.nonempty.strings}
\title{Tests whether 'x' is a character vector of nonempty strings.}
\usage{
is.vector.of.nonempty.strings(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE if and only if 'x' is a vector of nonempty strings none of
  which is an NA, otherwise FALSE. Empty vectors yield FALSE.
}
\description{
Tests whether 'x' is a character vector of nonempty strings.
}

