% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostrata.R
\name{IsFixedRandomization}
\alias{IsFixedRandomization}
\title{Test if randomizing the geostrata can lead only to a single outcome.}
\usage{
IsFixedRandomization(geostrata)
}
\arguments{
\item{geostrata}{a GeoStrata object.}
}
\value{
\code{TRUE} if randomizing the geostrata can only lead to a single
\code{GeoAssignment}, \code{FALSE} otherwise.
}
\description{
Test if randomizing the geostrata can lead only to a single outcome.
}

