% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isinperiod.R
\name{IsInPeriod.TBRAnalysisFitTbr1}
\alias{IsInPeriod.TBRAnalysisFitTbr1}
\title{[internal] Returns a logical vector indicating which of the rows are in
the specified period(s) in the TBRAnalysisData object.}
\usage{
\method{IsInPeriod}{TBRAnalysisFitTbr1}(obj, periods)
}
\arguments{
\item{obj}{a TBRAnalysisFitTbr1 object.}

\item{periods}{(vector of strings) names of the periods.}
}
\value{
A logical vector of length \code{nrow(obj)}, TRUE for a row that is
  within the given period(s), FALSE otherwise. No NAs.
}
\description{
[internal] Returns a logical vector indicating which of the rows are in
the specified period(s) in the TBRAnalysisData object.
}

