% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostrata.R
\name{IsStratumCompatibleWithRatios}
\alias{IsStratumCompatibleWithRatios}
\title{Checks whether strata are compatible with group.ratios.}
\usage{
IsStratumCompatibleWithRatios(geo.group, group.ratios)
}
\arguments{
\item{geo.group}{(integer vector of length equal to number of geos) a vector
of geo group numbers (may be NAs, but zeros are not allowed).}

\item{group.ratios}{(integer vector of length equal to number of geo groups)
vector of ratios of the sizes of each group.}
}
\value{
TRUE if geo.group is compatible with group.ratios, i.e. no group
  has been assigned more geos than what group.ratios allows. FALSE
  otherwise.
}
\description{
Checks whether strata are compatible with group.ratios.
}

