% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgeogroups.R
\name{MapGeoGroups<-.GeoExperimentData}
\alias{MapGeoGroups<-.GeoExperimentData}
\title{Map or merge geo group numbers into new group ids.}
\usage{
\method{MapGeoGroups}{GeoExperimentData}(obj) <- value
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{value}{(integer vector or NA) mapping of old group numbers to new
ones. See 'MapGeoGroups<-.GeoAssignment' for details.}
}
\value{
The GeoExperimentData object, with the column 'geo.group' affected.
}
\description{
Map or merge geo group numbers into new group ids.
}

