% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapgeogroups.R
\name{MapGeoGroups<-}
\alias{MapGeoGroups<-}
\title{Map or merge geo group numbers into new group ids.}
\usage{
MapGeoGroups(obj) <- value
}
\arguments{
\item{obj}{an object.}

\item{value}{(integer vector, NA allowed, NULL or empty integer vector
allowed) mapping of old group numbers to new ones. The length must be
exactly equal to the number of old groups in the object. The old
numbers match the positions of the index, and the new ones are the
values in the vector. For example, c(2, 1) maps 1->2 and 2->1; c(3,
2, 1) maps 1->3, 3->1 and 2 stays unchanged; '1:2' has no effect;
NOTE: c(1, 1) merges groups 1 and 2 into 1! If the 'geo.group' is NA
in the object, its value will be NOT changed. If the data has only
NAs, the only mapping that is allowed is an empty vector, resulting
in no change in the original object. The special value '0' (group of
excluded geos) will also be unchanged.}
}
\value{
An object identical to 'obj' except the 'geo.groups' column has
  (possibly) changed.
}
\description{
Map or merge geo group numbers into new group ids.
}

