% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_methods.R
\name{merge.GeoExperimentData}
\alias{merge.GeoExperimentData}
\title{Merges two GeoExperimentData objects.}
\usage{
\method{merge}{GeoExperimentData}(x, y, ...)
}
\arguments{
\item{x}{a GeoExperimentData object.}

\item{y}{a GeoExperimentData object.}

\item{...}{ignored.}
}
\value{
A GeoExperimentData object.
}
\description{
Merges two GeoExperimentData objects.
}
\note{
The object created is a GeoExperimentData that includes all the
columns of x, and all the columns of y that were not in x. Columns of y
that have the same name as columns of x are ignored and a warning is
issued whenever y and x share some column name(s). The merging is done
by kDate, kGeo, kPeriod, kGeoGroup, kAssignment.
}

