% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_methods.R
\name{merge.GeoTimeseries}
\alias{merge.GeoTimeseries}
\title{Merges two GeoTimeseries objects.}
\usage{
\method{merge}{GeoTimeseries}(x, y, ...)
}
\arguments{
\item{x}{a GeoTimeseries object.}

\item{y}{a GeoTimeseries object.}

\item{...}{ignored.}
}
\value{
A GeoTimeseries object.
}
\description{
Merges two GeoTimeseries objects.
}
\note{
The object created is a GeoTimeseries that includes all the columns
of x, and all the columns of y that were not in x. Columns of y that
have the same name as columns of x are ignored and a warning is issued
whenever y and x share some column name(s). The merging is done by
kDate, kGeo.
}

