% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messaging.R
\name{Message}
\alias{Message}
\title{Form a message, prefixed with the message context string.}
\usage{
Message(...)
}
\arguments{
\item{...}{R vectors, usually character, to be collapsed using 'paste0'.}
}
\value{
A string.
}
\description{
Form a message, prefixed with the message context string.
}

