% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messaging.R
\name{Messagef}
\alias{Messagef}
\title{Form a message, using sprintf.}
\usage{
Messagef(fmt, ...)
}
\arguments{
\item{fmt}{a character vector of format strings (see 'sprintf' for details).}

\item{...}{values to be passed to 'fmt'.}
}
\value{
A string.
}
\description{
Form a message, using sprintf.
}

