% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.Geos}
\alias{plot.Geos}
\title{Generate a (gg)plot of a Geos object.}
\usage{
\method{plot}{Geos}(x, y = NULL, geom = c("bar", "point", "rect"),
  sort.by = y, log.scale = FALSE, ...)
}
\arguments{
\item{x}{a Geos object.}

\item{y}{(string) name of the metric to plot. If omitted, plots the volume
defined in the object by default.}

\item{geom}{(string) short name of the geom ggplot2 method to be used.}

\item{sort.by}{(string) name of the column to sort the plot by, defaults to
y. If the column contains string, alphabetical order will be used
unless every string can be coerced to an integer in which case
numerical (ascending) order will be used. If the column contains
numerical values, numerical (descending) order will be used.}

\item{log.scale}{(flag) plot y-axis on log scale?}

\item{...}{ignored.}
}
\value{
A ggplot object.
}
\description{
Generate a (gg)plot of a Geos object.
}

