% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.GeoTimeseries}
\alias{plot.GeoTimeseries}
\title{Generate a (gg)plot of a GeoTimeseries object.}
\usage{
\method{plot}{GeoTimeseries}(x, y = NULL, by = NULL, subset = NULL,
  aggregate = FALSE, title = y, log.scale = FALSE, legend = TRUE, ...)
}
\arguments{
\item{x}{a GeoTimeseries object.}

\item{y}{(string) name of the metric to plot. If omitted, plots the first
metric in the object by default.}

\item{by}{(string) group by which column? Default is 'geo'.}

\item{subset}{(vector) subset of the items in column 'by' to use.}

\item{aggregate}{(flag) aggregate over?}

\item{title}{(string) a title string. By default, the name of the metric to
plot.}

\item{log.scale}{(flag) plot y-axis on log scale?}

\item{legend}{(flag) plot the legend?}

\item{...}{ignored.}
}
\value{
A ggplot object.
}
\description{
Generate a (gg)plot of a GeoTimeseries object.
}
\note{
Sets the lower y-axis limit to zero if all 'y' values are nonnegative
and log.scale=\code{FALSE}.
}

