% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tbrplotdata.R
\name{plot.TBRAnalysisFitTbr1}
\alias{plot.TBRAnalysisFitTbr1}
\title{Creates a graph of the TBR analysis fit.}
\usage{
\method{plot}{TBRAnalysisFitTbr1}(x, y, panels = GetTBRPlotPanelNames(),
  periods = c("pretest", "prediction"), quantiles = c(0.1, 0.9),
  highlight.weeks = TRUE, date.format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{a TBRAnalysisFitTbr1 object.}

\item{y}{ignored.}

\item{panels}{(vector of strings) names of the panels to be plotted.}

\item{periods}{(vector of strings) names of the periods to show.}

\item{quantiles}{(real vector of length 2) lower and upper quantiles of the
credible interval to show.}

\item{highlight.weeks}{(flag), alternate background shading to highlight
weeks?}

\item{date.format}{(string) date format.}

\item{...}{further arguments passed to methods 'as.TBRPlotData' and
'plot.TBRPlotData'.}
}
\value{
A ggplot2 object.
}
\description{
Creates a graph of the TBR analysis fit.
}

