% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tbrplotdata.R
\name{plot.TBRPlotData}
\alias{plot.TBRPlotData}
\title{Creates a graph of the TBRPlotData object.}
\usage{
\method{plot}{TBRPlotData}(x, y, highlight.weeks = TRUE,
  date.format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{a TBRPlotData object.}

\item{y}{ignored.}

\item{highlight.weeks}{(flag), alternate background shading to highlight
weeks?}

\item{date.format}{(string) date format.}

\item{...}{ignored.}
}
\value{
A ggplot2 object.
}
\description{
Creates a graph of the TBRPlotData object.
}

