% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tbrplotdata.R
\name{plot.TBRROASAnalysisFit}
\alias{plot.TBRROASAnalysisFit}
\title{Creates a graph of the TBR ROAS analysis fit.}
\usage{
\method{plot}{TBRROASAnalysisFit}(x, y, panels = "cumulative",
  periods = "prediction", quantiles = c(0.1, 0.9), highlight.weeks = TRUE,
  date.format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{a TBRROASAnalysisFit object.}

\item{y}{ignored.}

\item{panels}{(vector of strings) names of the panels to be plotted.}

\item{periods}{(vector of strings) names of the periods to show.}

\item{quantiles}{(real vector of length 2) lower and upper quantiles of the
credible interval to show.}

\item{highlight.weeks}{(flag), alternate background shading to highlight
weeks?}

\item{date.format}{(string) date format.}

\item{...}{further arguments passed to methods \code{\link{as.TBRPlotData}} and
\code{\link{plot.TBRPlotData}}.}
}
\value{
A ggplot2 object.
}
\description{
Creates a graph of the TBR ROAS analysis fit.
}

