% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbrroasanalysisfit_gbr1.R
\name{print.GBRROASAnalysisFitGbr1}
\alias{print.GBRROASAnalysisFitGbr1}
\title{Shows a summary of GBR analysis results.}
\usage{
\method{print}{GBRROASAnalysisFitGbr1}(x, ...)
}
\arguments{
\item{x}{a GBRROASAnalysisFitGbr1 object.}

\item{...}{ignored.}
}
\value{
The object itself, invisibly. As a side effect, prints the default
  summary of 'x' on the console.
}
\description{
Shows a summary of GBR analysis results.
}

