% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateroasanalysis.R
\name{print.ROASPreanalysisFit}
\alias{print.ROASPreanalysisFit}
\title{Prints a summary of the \code{ROASPreanalysisFit} method.}
\usage{
\method{print}{ROASPreanalysisFit}(x, ...)
}
\arguments{
\item{x}{a \code{ROASPreanalysisFit} object.}

\item{...}{arguments passed to the 'summary' method.}
}
\value{
Prints the summary of \code{x} and returns the summary of \code{x}
invisibly.
}
\description{
Prints a summary of the \code{ROASPreanalysisFit} method.
}

