% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbrroasanalysis.R
\name{print.TBRROASAnalysisFit}
\alias{print.TBRROASAnalysisFit}
\title{Shows a summary of TBR analysis results.}
\usage{
\method{print}{TBRROASAnalysisFit}(x, ...)
}
\arguments{
\item{x}{a TBRROASAnalysisFit object.}

\item{...}{ignored.}
}
\value{
The object itself, invisibly. As a side effect, prints the default
  summary of 'x' on the console.
}
\description{
Shows a summary of TBR analysis results.
}

