% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.TBRAnalysisFitTbr1}
\alias{quantile.TBRAnalysisFitTbr1}
\title{Computes quantiles of the cumulative, pointwise, or the counterfactual
(posterior) distribution of the causal lift, from the beginning of
pretest to the end of posttest.}
\usage{
\method{quantile}{TBRAnalysisFitTbr1}(x, probs = c(0.1, 0.5, 0.9),
  distribution = c("cumulative", "pointwise", "counterfactual"), ...)
}
\arguments{
\item{x}{a TBRAnalysisFitTbr1 object.}

\item{probs}{(numeric vector, each >= 0 and <= 1) quantiles to calculate.}

\item{distribution}{(string) either 'cumulative' or 'pointwise' or
'counterfactual'.}

\item{...}{ignored.}
}
\value{
A TBRQuantiles object, which is a data frame with each row
  corresponding to a day in the pretest and the prediction period, with
  columns 'date', 'test', and the columns for the quantiles.
}
\description{
Computes quantiles of the cumulative, pointwise, or the counterfactual
(posterior) distribution of the causal lift, from the beginning of
pretest to the end of posttest.
}

