% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.TBRROASAnalysisFit}
\alias{quantile.TBRROASAnalysisFit}
\title{Computes quantiles of the cumulative or pointwise distribution of the
incremental ROAS, for each of the days in the pretest and the
prediction period.}
\usage{
\method{quantile}{TBRROASAnalysisFit}(x, probs = c(0.1, 0.5, 0.9),
  distribution = c("cumulative", "pointwise"), ...)
}
\arguments{
\item{x}{a TBRROASAnalysisFit object.}

\item{probs}{(numeric vector, each >= 0 and <= 1) quantiles to calculate.}

\item{distribution}{(string) either 'cumulative' or 'pointwise'.}

\item{...}{ignored.}
}
\value{
A TBRQuantiles object, which is a data frame with each row
  corresponding to a day in the pretest and the prediction period, with
  columns 'date', 'test', and the columns for the quantiles.
}
\description{
Computes quantiles of the cumulative or pointwise distribution of the
incremental ROAS, for each of the days in the pretest and the
prediction period.
}
\note{
The pretest period is included solely for the purpose of obtaining
objects of conforming size from both quantile methods (that of
TBRAnalysisFitTbr1 and of this one).
}

