% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{Randomize.Geos}
\alias{Randomize.Geos}
\title{Randomize geos, assigning geo groups by strata.}
\usage{
\method{Randomize}{Geos}(obj, n.groups = 2, group.ratios = rep(1, length.out
  = n.groups), ...)
}
\arguments{
\item{obj}{a Geos object.}

\item{n.groups}{(integer) number of groups.}

\item{group.ratios}{(integer vector of length n.groups) vector of ratios of
the sizes of each group. By default each group is assumed to have
equal ratios.}

\item{...}{ignored.}
}
\value{
A GeoAssignment object.
}
\description{
Randomize geos, assigning geo groups by strata.
}

