% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{Randomize.GeoStrata}
\alias{Randomize.GeoStrata}
\title{Randomize geos, assigning geo groups by strata.}
\usage{
\method{Randomize}{GeoStrata}(obj, ...)
}
\arguments{
\item{obj}{a GeoStrata object.}

\item{...}{ignored.}
}
\value{
A GeoAssignment object.
}
\description{
Randomize geos, assigning geo groups by strata.
}

