% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{RenameColumns}
\alias{RenameColumns}
\title{Renames columns of a data frame.}
\usage{
RenameColumns(x, map = character(0))
}
\arguments{
\item{x}{a data frame.}

\item{map}{a named character vector, mapping old column names to new ones
such that the new ones are in 'names' and the values are the old
ones. For example, \code{c(geo='GMA', date='Week Ending')}. If 'map'
has length 0, the original data frame is returned.}
}
\value{
The data frame, with the column names renamed.
}
\description{
Renames columns of a data frame.
}

