% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateroasanalysis.R
\name{.ROASPreanalysisFitRow}
\alias{.ROASPreanalysisFitRow}
\title{[internal] Constructs a \code{ROASPreanalysisFit}-compatible object with one
single row. To be bound together row-wise with other such objects.}
\usage{
.ROASPreanalysisFitRow(sim, model, estimate, sd, df, i, geo.assignment)
}
\arguments{
\item{sim}{(integer >= 0) number of the simulation.}

\item{model}{(string) model id.}

\item{estimate}{(real number) point estimate.}

\item{sd}{(real number > 0) posterior s.d.}

\item{df}{(integer > 0) degrees of freedom of the t distribution that is the
posterior of the iROAS estimate.}

\item{i}{(integer >= 1) index of the data set.}

\item{geo.assignment}{(a \code{GeoAssignment} object) the actual geo
assignment used.}
}
\value{
A \code{ROASPreanalysisFitRow} object.
}
\description{
[internal] Constructs a \code{ROASPreanalysisFit}-compatible object with one
single row. To be bound together row-wise with other such objects.
}
\note{
No integrity checking of the arguments is done.
}

