% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roasanalysisresults.R
\name{round.ROASAnalysisResults}
\alias{round.ROASAnalysisResults}
\title{Rounds the estimates to a given number of decimal places.}
\usage{
\method{round}{ROASAnalysisResults}(x, digits = 1)
}
\arguments{
\item{x}{a ROASAnalysisResults object.}

\item{digits}{number of digital places to round to.}
}
\value{
A ROASAnalysisResults object, with estimates appropriately rounded.
}
\description{
Rounds the estimates to a given number of decimal places.
}

