% Copyright (C) 2017 Google, Inc.

\name{salesandcost}
\docType{data}
\alias{salesandcost}
\title{Sales and Cost Data}
\usage{data(salesandcost)}
\description{
  A data frame with 9225 rows, with the columns:
  \describe{
    \item{date:}{A character string representing a date in the format
      MM/DD/YYYY.}
    \item{geo:}{Geo id, integer.}
    \item{sales:}{Sales for the corresponding geo and day, numeric.}
    \item{cost:}{Cost of advertising, numeric.}
  }
}
\source{
Simulated data.
}
