% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setexperimentperiods.R
\name{SetExperimentPeriods<-.GeoExperimentData}
\alias{SetExperimentPeriods<-.GeoExperimentData}
\title{Associates the object with an ExperimentPeriods object, mapping the 'date'
column to a 'period' column.}
\usage{
\method{SetExperimentPeriods}{GeoExperimentData}(obj, strict = TRUE, ...) <-
  value
}
\arguments{
\item{obj}{the object to change.}

\item{strict}{(flag) insist that data has all specified experiment periods?}

\item{...}{ignored.}

\item{value}{a ExperimentPeriods object.}
}
\value{
The same object that has been modified in place; the 'period' slot
  is assigned with the new value and the columns 'period' and
  'assignment' are changed to reflect the new date ranges.
}
\description{
Associates the object with an ExperimentPeriods object, mapping the 'date'
column to a 'period' column.
}
\seealso{
SetExperimentPeriods<- (generic).
}

