% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setexperimentperiods.R
\name{SetExperimentPeriods<-}
\alias{SetExperimentPeriods<-}
\title{Associates the object with an ExperimentPeriods object.}
\usage{
SetExperimentPeriods(obj, ...) <- value
}
\arguments{
\item{obj}{the object to change.}

\item{...}{further arguments passed to methods.}

\item{value}{a ExperimentPeriods object.}
}
\value{
The object that has been modified in place.
}
\description{
Associates the object with an ExperimentPeriods object.
}
\seealso{
SetExperimentPeriods<-.GeoExperimentData.
}

