% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setexperimentperiods.R
\name{SetExperimentPeriods<-}
\alias{SetExperimentPeriods<-}
\alias{SetExperimentPeriods<-.GeoExperimentData}
\title{Associates the object with an ExperimentPeriods object.}
\usage{
SetExperimentPeriods(obj, ...) <- value

\method{SetExperimentPeriods}{GeoExperimentData}(obj, strict = TRUE, ...) <-
  value
}
\arguments{
\item{obj}{the object to change.}

\item{...}{further arguments passed to methods.}

\item{value}{a ExperimentPeriods object.}

\item{strict}{(flag) insist that data has all specified experiment periods?}
}
\value{
The object that has been modified in place.
}
\description{
Associates the object with an ExperimentPeriods object.
}
\note{
The 'date' column is mapped to the 'period' column. The 'period' slot
is assigned with the new value and the columns 'period' and 'assignment' are
changed to reflect the new date ranges.
}

