% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeoassignment.R
\name{SetGeoAssignment<-.GeoExperimentData}
\alias{SetGeoAssignment<-.GeoExperimentData}
\title{Associates the object with a GeoAssignment object, mapping geos into geo
group numbers.}
\usage{
\method{SetGeoAssignment}{GeoExperimentData}(obj, strict = TRUE, ...) <- value
}
\arguments{
\item{obj}{the GeoExperimentData object to change.}

\item{strict}{(flag) insist that all geos in the data are mapped? Also, warn
if some geos are not found in the data?}

\item{...}{ignored.}

\item{value}{a GeoAssignment object to associate with the GeoExperimentData
object.}
}
\value{
The same object that has been modified in place; the
  'geo.assignment' slot is assigned with the new value and the columns
  'geo.group' and 'assignment' are changed to reflect the new geo
  assignment.
}
\description{
Associates the object with a GeoAssignment object, mapping geos into geo
group numbers.
}
\note{
If 'value' is 'NULL', the geo assignment and the treatment assignment
are removed, and their corresponding columns in the data frame are
reset to 'NA'. An error is thrown if any geo in data cannot be mapped
to a group (due to the GeoAssignment object not having such a mapping).
Setting 'strict' to FALSE will avert this. An error is thrown if some
of the geos in the mapping were not present in the data. Setting
'strict' to FALSE will avert this.
}
\seealso{
SetGeoAssignment<- (generic).
}

