% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeoassignment.R
\name{SetGeoAssignment<-}
\alias{SetGeoAssignment<-}
\title{Associates the object with a GeoAssignment object.}
\usage{
SetGeoAssignment(obj, ...) <- value
}
\arguments{
\item{obj}{the object to change.}

\item{...}{further arguments passed to methods.}

\item{value}{a GeoAssignment object.}
}
\value{
The object (that has been modified in place).
}
\description{
Associates the object with a GeoAssignment object.
}
\seealso{
SetGeoAssignment.GeoExperimentData.
}

