% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeoassignment.R
\name{SetGeoAssignment<-}
\alias{SetGeoAssignment<-}
\alias{SetGeoAssignment<-.GeoExperimentData}
\title{Associates the object with a \code{GeoAssignment} object.}
\usage{
SetGeoAssignment(obj, ...) <- value

\method{SetGeoAssignment}{GeoExperimentData}(obj, strict = TRUE, ...) <- value
}
\arguments{
\item{obj}{the object to change.}

\item{...}{further arguments passed to methods.}

\item{value}{a \code{GeoAssignment} object.}

\item{strict}{(flag) insist that all geos in the data are mapped? Also, warn
if some geos are not found in the data?}
}
\value{
The object (that has been modified in place).
}
\description{
Associates the object with a \code{GeoAssignment} object.

Associates the object with a GeoAssignment object, mapping geos into geo
group numbers.
}
\details{
If \code{value} is \code{NULL}, the geo assignment and the treatment
assignment are removed, and their corresponding columns in the data frame
are reset to 'NA'. An error is thrown if any geo in data cannot be mapped to
a group (due to the \code{GeoAssignment} object not having such a mapping).
Setting 'strict' to \code{FALSE} will avert this. An error is thrown if some
of the geos in the mapping were not present in the data. Setting 'strict' to
\code{FALSE} will avert this.

\code{geo.assignment} slot is assigned with the new value and the columns
\code{geo.group} and \code{assignment} are changed to reflect the new geo
assignment.
}

