% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeogroup.R
\name{SetGeoGroup<-.GeoAssignment}
\alias{SetGeoGroup<-.GeoAssignment}
\title{Modifies a GeoAssignment object.}
\usage{
\method{SetGeoGroup}{GeoAssignment}(obj) <- value
}
\arguments{
\item{obj}{a GeoAssignment object.}

\item{value}{a GeoAssignment object.}
}
\value{
The same object, with the geo.group column modified.
}
\description{
Modifies a GeoAssignment object.
}

