% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeogroup.R
\name{SetGeoGroup<-.GeoExperimentData}
\alias{SetGeoGroup<-.GeoExperimentData}
\title{Modifies the GeoAssignment object in the GeoExperimentData object.}
\usage{
\method{SetGeoGroup}{GeoExperimentData}(obj) <- value
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{value}{a GeoAssignment object.}
}
\value{
The same object, with the geo assignment (and therefore also the
  geo.group column) modified.
}
\description{
Modifies the GeoAssignment object in the GeoExperimentData object.
}
\note{
The GeoExperimentData object must contain a valid GeoAssignment
object. Use SetGeoAssignment<- to associate a GeoAssignment object to
the GeoExperimentData object.
}

