% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeogroup.R
\name{SetGeoGroup<-.GeoStrata}
\alias{SetGeoGroup<-.GeoStrata}
\title{Modifies the geo assignment in a GeoStrata object.}
\usage{
\method{SetGeoGroup}{GeoStrata}(obj) <- value
}
\arguments{
\item{obj}{a GeoStrata object.}

\item{value}{a GeoAssignment object.}
}
\value{
The same object, with the geo.group column modified.
}
\description{
Modifies the geo assignment in a GeoStrata object.
}

