% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeogroup.R
\name{SetGeoGroup<-}
\alias{SetGeoGroup<-}
\title{Modifies the geo assignment in an object.}
\usage{
SetGeoGroup(obj) <- value
}
\arguments{
\item{obj}{an object with the columns 'geo' and 'geo.group'.}

\item{value}{a GeoAssignment object.}
}
\value{
The object with the modfied geo-to-geo.group mapping.
}
\description{
Modifies the geo assignment in an object.
}
\note{
See also: SetGeoAssignment<-.
}

