% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setgeogroup.R
\name{SetGeoGroup<-}
\alias{SetGeoGroup}
\alias{SetGeoGroup<-}
\alias{SetGeoGroup<-.GeoAssignment}
\alias{SetGeoGroup<-.GeoExperimentData}
\alias{SetGeoGroup<-.GeoStrata}
\title{Modifies the geo assignment in an object.}
\usage{
SetGeoGroup(obj) <- value

\method{SetGeoGroup}{GeoStrata}(obj) <- value

\method{SetGeoGroup}{GeoAssignment}(obj) <- value

\method{SetGeoGroup}{GeoExperimentData}(obj) <- value
}
\arguments{
\item{obj}{an object with the columns 'geo' and 'geo.group'.}

\item{value}{a GeoAssignment object.}
}
\value{
The object with the modfied geo-to-geo.group mapping.
}
\description{
Modifies the geo assignment in an object.
}
\note{
The \code{\link{GeoExperimentData}} object must contain a valid
\code{\link{GeoAssignment}} object. (Use \code{\link{SetGeoAssignment<-}} to
associate a \code{\link{GeoAssignment}} object to the
\code{\link{GeoExperimentData}} object.)
}
\seealso{
\code{\link{SetGeoAssignment<-}}, \code{\link{MapGeoGroups<-}}.
}

