% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setincrementalresponse.R
\name{SetIncrementalResponse<-.GeoExperimentData}
\alias{SetIncrementalResponse<-.GeoExperimentData}
\title{Sets an incremental response for simulation purposes in a
GeoExperimentData object.}
\usage{
\method{SetIncrementalResponse}{GeoExperimentData}(obj, response,
  periods = "all") <- value
}
\arguments{
\item{obj}{a GeoExperimentData object with the column '.spend'.}

\item{response}{(string) name of the column which acts as the response
metric.}

\item{periods}{numbers of the periods which are affected; if this is "all",
all periods are affected.}

\item{value}{(a real number) the incremental return on ad spend. Can be zero
or negative. Can also be NA, in the case of which the column
'.response' will be reset to NA.}
}
\value{
The same object, with the column '.response' modified.
}
\description{
Sets an incremental response for simulation purposes in a
GeoExperimentData object.
}
\note{
Creates a column '.response' if it does not exist, and registers it
as a metric. The column '.response' contains the baseline response and
the incremental response, which is defined as the specified incremental
ROAS ('value') times the column '.spend'.
}

