% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setincrementalresponse.R
\name{SetIncrementalResponse<-}
\alias{SetIncrementalResponse<-}
\title{Sets an incremental response for simulation purposes.}
\usage{
SetIncrementalResponse(obj, response, periods = "all") <- value
}
\arguments{
\item{obj}{an object with the column '.spend'.}

\item{response}{(string) name of the column which acts as the response
metric.}

\item{periods}{numbers of the periods which are affected; if this is "all",
all periods are affected.}

\item{value}{the absolute value of the incremental return on ad spend. Can
also be negative. Can also be NA, in the case of which the column
'.response' will be reset to NA.}
}
\value{
The object, with the modfied spend change column '.response'.
}
\description{
Sets an incremental response for simulation purposes.
}

