% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_class.R
\name{SetInfo}
\alias{SetInfo}
\title{Set the values of the 'info' attribute of an object.}
\usage{
SetInfo(obj, ..., info = NULL)
}
\arguments{
\item{obj}{some object.}

\item{...}{name-value pairs to set in the list-valued attribute 'info'.}

\item{info}{the initial value of the 'info' attribute; if specified, will
re-set the value before changing the individual name-value pairs.}
}
\value{
The object ('obj') that was changed.
}
\description{
Set the values of the 'info' attribute of an object.
}
\note{
Used only internally in this package. Do not use for data tables.
}

