% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messaging.R
\name{SetMessageContextString}
\alias{SetMessageContextString}
\title{Save a string representing the current context of the program flow.}
\usage{
SetMessageContextString(context)
}
\arguments{
\item{context}{(a string) context of the message; to be displayed in (error)
messages.}
}
\value{
The previous message context strings (a character vector).
}
\description{
Save a string representing the current context of the program flow.
}

