% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspendchange.R
\name{SetSpendChange<-.GeoExperimentData}
\alias{SetSpendChange<-.GeoExperimentData}
\title{Sets the spend change column in a GeoExperimentData object.}
\usage{
\method{SetSpendChange}{GeoExperimentData}(obj, prop.to, periods = "all") <-
  value
}
\arguments{
\item{obj}{a GeoExperimentData object. The values in the 'assignment' column
must not be missing.}

\item{prop.to}{(existing) name of the column in proportion to which the
spend change will be spread across the geos.}

\item{periods}{numbers of the periods which are affected; if this is "all",
all periods are affected and the spend change will be spread across
all periods that have spend change (determined by the column
'assignment').}

\item{value}{the total absolute value of the spend change. Can also be NA,
in the case of which the spend change column will be reset to NA.}
}
\value{
The same object, with the column specified as the spend change
  column '.spend' modified.
}
\description{
Sets the spend change column in a GeoExperimentData object.
}
\note{
Creates a column '.spend'. The spend change type is determined by the
'assignment' column. The absolute value of the spend change will be
spread across geos in the proportion of the given column ('prop.to');
the sign of the change is determined by the type of change: for
treatment assignments set to 'decrease', the sign will be negative,
otherwise positive.
}

