% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspendchange.R
\name{SetSpendChange<-}
\alias{SetSpendChange<-}
\title{Sets the spend change in a geo experiment for simulation purposes.}
\usage{
SetSpendChange(obj, prop.to, periods = "all") <- value
}
\arguments{
\item{obj}{an object with the column 'assignment'.}

\item{prop.to}{(existing) name of the column in proportion to which the
spend change will be spread across the geos.}

\item{periods}{numbers of the periods which are affected; if this is "all",
all periods are affected and the spend change will be spread across
all periods that have spend change (determined by the column
'assignment').}

\item{value}{the total absolute value of the spend change. Can also be NA,
in the case of which the spend change column will be reset to NA.}
}
\value{
The object, with the modified spend change column \code{.spend}.
}
\description{
Sets the spend change in a geo experiment for simulation purposes.
}

