% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspendchange.R
\name{SetSpendChange<-}
\alias{SetSpendChange<-}
\alias{SetSpendChange<-.GeoExperimentData}
\title{Sets the spend change in a geo experiment for simulation purposes.}
\usage{
SetSpendChange(obj, prop.to, periods = "all") <- value

\method{SetSpendChange}{GeoExperimentData}(obj, prop.to, periods = "all") <-
  value
}
\arguments{
\item{obj}{an object with the column 'assignment'.}

\item{prop.to}{(existing) name of the column in proportion to which the
spend change will be spread across the geos.}

\item{periods}{numbers of the periods which are affected; if this is
\code{"all"}, all periods are affected and the spend change will be spread
across all periods that have spend change (determined by the column
\code{assignment}).}

\item{value}{the total absolute value of the spend change. Can also be NA,
in the case of which the spend change column will be reset to NA.}
}
\value{
The object, with the modified spend change column \code{.spend}.
}
\description{
Sets the spend change in a geo experiment for simulation purposes.
}
\note{
Creates a column \code{.spend}. The spend change type is determined by the
\code{assignment} column. The absolute value of the spend change will be
spread across geos in the proportion of the given column (\code{prop.to});
the sign of the change is determined by the type of change: for
treatment assignments set to 'decrease', the sign will be negative,
otherwise positive.
}

