% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settreatmentassignment.R
\name{SetTreatmentAssignment<-.GeoExperimentData}
\alias{SetTreatmentAssignment<-.GeoExperimentData}
\title{Associates the object with an TreatmentAssignment object, mapping the
('period', 'geo.group') pairs to a (treatment) 'assignment' column.}
\usage{
\method{SetTreatmentAssignment}{GeoExperimentData}(obj, strict = TRUE, ...) <-
  value
}
\arguments{
\item{obj}{the object (a GeoExperimentData object) to change.}

\item{strict}{(flag) insist that data has all treatment combinations?}

\item{...}{ignored.}

\item{value}{a TreatmentAssignment object.}
}
\value{
The object, that has been modified in place.
}
\description{
Associates the object with an TreatmentAssignment object, mapping the
('period', 'geo.group') pairs to a (treatment) 'assignment' column.
}
\note{
If 'strict' is TRUE, throws an error if some treatment assignments
are not found in the data. If 'strict' is FALSE, no warnings or errors
are thrown.
}
\seealso{
SetTreatmentAssignment<- (treatment).
}

