% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentassignment.R
\name{SetTreatmentAssignment<-}
\alias{SetTreatmentAssignment<-}
\title{Associates the object with an TreatmentAssignment object.}
\usage{
SetTreatmentAssignment(obj, ...) <- value
}
\arguments{
\item{obj}{the object to change.}

\item{...}{further arguments passed to methods.}

\item{value}{a TreatmentAssignment object.}
}
\value{
The object (that has been modified in place).
}
\description{
Associates the object with an TreatmentAssignment object.
}
\seealso{
SetTreatmentAssignment<-.GeoExperimentData.
}

