% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settreatmentassignment.R
\name{SetTreatmentAssignment<-}
\alias{SetTreatmentAssignment<-}
\alias{SetTreatmentAssignment<-.GeoExperimentData}
\title{Associates the object with an TreatmentAssignment object.}
\usage{
SetTreatmentAssignment(obj, ...) <- value

\method{SetTreatmentAssignment}{GeoExperimentData}(obj, strict = TRUE, ...) <-
  value
}
\arguments{
\item{obj}{the object to change.}

\item{...}{further arguments passed to methods.}

\item{value}{a TreatmentAssignment object.}

\item{strict}{(flag) insist that data has all treatment combinations?}
}
\value{
The object (that has been modified in place).
}
\description{
Associates the object with an TreatmentAssignment object.
}
\note{
The value pairs ('period', 'geo.group') are mapped to a (treatment)
'assignment' column.

If 'strict' is TRUE, throws an error if some treatment assignments
are not found in the data. If 'strict' is FALSE, no warnings or errors
are thrown.
}

