% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roasanalysisresults.R
\name{signif.ROASAnalysisResults}
\alias{signif.ROASAnalysisResults}
\title{Rounds the estimates to a given number of significant digits.}
\usage{
\method{signif}{ROASAnalysisResults}(x, digits = 1)
}
\arguments{
\item{x}{a \code{ROASAnalysisResults} object.}

\item{digits}{number of significant digits to round to.}
}
\value{
A \code{ROASAnalysisResults} object, with estimates appropriately
rounded.
}
\description{
Rounds the estimates to a given number of significant digits.
}

