% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulategeoxdataset.R
\name{SimulateGeoExperimentData.GeoExperimentPreanalysisData}
\alias{SimulateGeoExperimentData.GeoExperimentPreanalysisData}
\title{Returns a set number i from the GeoExperimentPreanalysisData object.}
\usage{
\method{SimulateGeoExperimentData}{GeoExperimentPreanalysisData}(obj,
  i = NA_integer_, ...)
}
\arguments{
\item{obj}{a GeoExperimentPreanalysisData object.}

\item{i}{(positive integer or NA) number of the pseudo data set; if NA, the
number will be drawn from a uniform discrete distribution.}

\item{...}{ignored.}
}
\value{
A GeoExperimentData object with the experiment periods, geo
  assignment, and the treatment assignment set.
}
\description{
Returns a set number i from the GeoExperimentPreanalysisData object.
}
\note{
If the embedded 'geos' object is GeoStrata, a randomization is done
and the geo assignment applied.
}

