% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateposterior.R
\name{SimulatePosterior}
\alias{SimulatePosterior}
\title{Draws a sample from a posterior distribution.}
\usage{
SimulatePosterior(obj, n.sims = 1e+05, ...)
}
\arguments{
\item{obj}{an object.}

\item{n.sims}{(integer >= 2) number of simulations to draw.}

\item{...}{other arguments passed on to the methods.}
}
\value{
A PosteriorSimulations object, which is a matrix with 'n.sims'
  columns. Each row corresponds to a variable.
}
\description{
Draws a sample from a posterior distribution.
}

