% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateroasanalysis.R
\name{.SimulateROASAnalysis}
\alias{.SimulateROASAnalysis}
\title{[internal] Performs a ROAS analysis on one simulated data set drawn from
the \code{GeoExperimentPreanalysisData} object.}
\usage{
.SimulateROASAnalysis(sim, randomize.i, obj, response, prop.to, cost = 1,
  models = GetModelIds(), swap.and.redo = FALSE)
}
\arguments{
\item{sim}{(integer >= 1) number of the simulation; if 'randomize.i' is
\code{FALSE}, 'sim' will be used as the number of the data set to draw.}

\item{randomize.i}{(flag) if TRUE, the number of the data set will be drawn
at random. If \code{FALSE}, the number of the data set will be set equal
to \code{sim}.}

\item{obj}{a \code{GeoExperimentPreanalysisData} object.}

\item{response}{(string) name of the metric column to use as the response
variable.}

\item{prop.to}{(string) an existing name of the column in proportion to
which the spend change will be distributed across the geos.}

\item{cost}{(real number) incremental cost, assumed fixed and known.}

\item{models}{(vector of nonempty strings) one or more analysis model ids.
The simulated data set is apply to simulated data set. By default,
all possible models are fit.}

\item{swap.and.redo}{(flag) swap groups and redo the analysis? Note: number
of groups must be 2 if this flag is TRUE. Since this produces two
simulations, the indices of simulations that are registered will be
\code{2 * sim - 1} and \code{2 * sim}.}
}
\value{
A \code{ROASAnalysisFit} object, a data frame with one or more
  rows.
}
\description{
[internal] Performs a ROAS analysis on one simulated data set drawn from
the \code{GeoExperimentPreanalysisData} object.
}
\note{
This is an internal function, called from another function. No
integrity checking of the parameters is done; this must be done by the
enclosing function.
}

