% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{Subset.GeoExperimentData}
\alias{Subset.GeoExperimentData}
\title{Extracts a subset of a GeoExperimentData.}
\usage{
\method{Subset}{GeoExperimentData}(obj, rows, columns, ...)
}
\arguments{
\item{obj}{a GeoExperimentData object.}

\item{rows}{logical vector indicating rows to keep. No missing values are
allowed.}

\item{columns}{character vector indicating columns of the data frame to
keep. Columns such as 'date', 'geo', 'period', 'geo.group' and
'assignment', if they exists, are always included.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A GeoExperimentData object (a subset of 'x').
}
\description{
Extracts a subset of a GeoExperimentData.
}
\note{
The arguments 'rows' and 'columns' are *not* evaluated in the
environment of the columns of 'obj'; the expressions must be evaluable
in the enclosing environment. This behavior is deliberately different
from that of 'subset'. The original experiment configuration (periods,
geo assignment, treatment assignment) is preserved.
}
\seealso{
Subset (generic), Subset.GeoTimeseries.
}

