% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{Subset.GeoTimeseries}
\alias{Subset.GeoTimeseries}
\title{Extract a subset of GeoTimeseries.}
\usage{
\method{Subset}{GeoTimeseries}(obj, rows, columns, ...)
}
\arguments{
\item{obj}{a GeoTimeseries object.}

\item{rows}{logical vector indicating rows to keep. No missing values are
allowed.}

\item{columns}{character vector indicating columns of the data frame to
keep. Columns such as 'date', 'geo', 'period', 'geo.group' and
'assignment', if they exist, are always included.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A GeoTimeseries object (a subset of 'obj').
}
\description{
Extract a subset of GeoTimeseries.
}
\note{
The arguments 'rows' and 'columns' are not evaluated in the
environment of the columns of 'obj'; the expressions must be evaluable
in the enclosing environment.
}
\seealso{
Subset (generic), Subset.GeoExperimentData.
}

