% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{Subset}
\alias{Subset}
\title{Returns a subset of the object, ensuring that the resulting object is of
the same class.}
\usage{
Subset(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of the same class.
}
\description{
Returns a subset of the object, ensuring that the resulting object is of
the same class.
}
\note{
A generic S3 method.
}
\seealso{
Subset.GeoTimeseries, Subset.GeoExperimentData.
}

