% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbrroasanalysisfit_gbr1.R
\name{summary.GBRROASAnalysisFitGbr1}
\alias{summary.GBRROASAnalysisFitGbr1}
\title{Returns a concise summary of the ROAS estimate and confidence interval and
the total incremental cost and incremental response.}
\usage{
\method{summary}{GBRROASAnalysisFitGbr1}(object, level = 0.9,
  interval.type = c("one-sided", "two-sided"), threshold = 0, ...)
}
\arguments{
\item{object}{a GBRROASAnalysisFitGbr1 object.}

\item{level}{(number between 0 and 1) confidence level.}

\item{interval.type}{(string) 'one-sided', or 'two-sided' (interval).}

\item{threshold}{(numeric vector) threshold(s) for the right-tail posterior
probabilities of beta2.}

\item{...}{ignored.}
}
\value{
A ROASAnalysisResults object.
}
\description{
Returns a concise summary of the ROAS estimate and confidence interval and
the total incremental cost and incremental response.
}

