% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doroaspreanalysis.R
\name{summary.ROASPreanalysisFit}
\alias{summary.ROASPreanalysisFit}
\title{Computes a summary of the predicted ROAS estimate and associated
incremental cost.}
\usage{
\method{summary}{ROASPreanalysisFit}(object, level = 0.9,
  interval.type = c("one-sided", "two-sided"), precision = 1,
  cost = NA_real_, ...)
}
\arguments{
\item{object}{a ROASPreanalysisFit object.}

\item{level}{(number between 0 and 1) confidence level.}

\item{interval.type}{(string) 'one-sided', 'two-sided'.}

\item{precision}{(number) target precision of the estimate; the distance
between the lower bound of the confidence interval and the point
estimate. Note: if 'cost' is given, this will be ignored and the CI
half-width will be computed based on 'cost'.}

\item{cost}{(number) cost difference (ad spend difference); if missing, will
be computed based on 'precision'.}

\item{...}{ignored.}
}
\value{
A ROASPreanalysisResults object.
}
\description{
Computes a summary of the predicted ROAS estimate and associated
incremental cost.
}

