% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tbranalysisfit_tbr1.R
\name{summary.TBRAnalysisFitTbr1}
\alias{summary.TBRAnalysisFitTbr1}
\title{Returns a concise summary of the incremental response estimate and its
confidence interval for the model 'tbr1'.}
\usage{
\method{summary}{TBRAnalysisFitTbr1}(object, level = 0.9,
  interval.type = c("one-sided", "two-sided"), threshold = 0, ...)
}
\arguments{
\item{object}{a \code{TBRAnalysisFit} object.}

\item{level}{(number between 0 and 1) confidence level.}

\item{interval.type}{(string) 'one-sided', 'two-sided' (interval).}

\item{threshold}{(numeric vector) threshold(s) for the right-tail posterior.}

\item{...}{ignored.}
}
\value{
A \code{TBRAnalysisResults} object.
}
\description{
Returns a concise summary of the incremental response estimate and its
confidence interval for the model 'tbr1'.
}

