% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbrroasanalysis.R
\name{summary.TBRROASAnalysisFit}
\alias{summary.TBRROASAnalysisFit}
\title{Returns a concise summary of the incremental response estimate and its
confidence interval for a TBR ROAS analysis.}
\usage{
\method{summary}{TBRROASAnalysisFit}(object, level = 0.9,
  interval.type = c("one-sided", "two-sided"), threshold = 0, ...)
}
\arguments{
\item{object}{a TBRROASAnalysisFit object.}

\item{level}{(number between 0 and 1) confidence level.}

\item{interval.type}{(string) 'one-sided', 'two-sided' (interval).}

\item{threshold}{(numeric vector) threshold(s) for the right-tail posterior
probabilities of the total cumulative iROAS.}

\item{...}{ignored.}
}
\value{
A ROASAnalysisResults object.
}
\description{
Returns a concise summary of the incremental response estimate and its
confidence interval for a TBR ROAS analysis.
}

