% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbranalysisdata.R
\name{TBRAnalysisData}
\alias{TBRAnalysisData}
\title{Constructs a TBRAnalysisData object.}
\usage{
TBRAnalysisData(x)
}
\arguments{
\item{x}{a _plain_ \code{data.frame} or an object that has to be coercible
to a _plain_ \code{data.frame} with the required columns \code{data}, which
_must_ be a \code{Date} vector; integer-valued column \code{period}
indicating the pre-experiment, pretest, intervention, cooldown, and posttest
periods; numeric columns \code{y} (response), \code{x}
(covariate). Optionally the data frame can have other columns, which are
ignored but whose names cannot start with a dot ('.'). No missing values are
allowed in the pretest and test periods (but allowed outside of either
period).#'}
}
\value{
A \code{TBRAnalysisData} object, which is similarly a
\code{data.frame} with the columns:

\code{TBRAnalysisData} is a \code{data.frame}, with the required columns,
\itemize{
  \item{\code{date}}: a 'Date'-valued vector.
  \item{\code{period}}: numeric indicator of various periods.
    \code{0} indicates the pretest period. \code{NA} indicates a date
    that has been excluded from the analyses.
  \item{\code{y}}: numeric values of the metric of the Treatment group.
  \item{\code{x}}: numeric values of the metric in the Control group.
    In addition, optionally any other user-definable columns.
}
}
\description{
Constructs a TBRAnalysisData object.
}
\note{
The column \code{date} is the primary key; it is guaranteed that no
rows with duplicate geos exist and that the rows are in temporal order.
}

