% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_objects.R
\name{assert_tests}
\alias{assert_tests}
\alias{is.character.vector}
\alias{is.empty.string}
\alias{is.inf}
\alias{is.integer.valued}
\alias{is.logical.vector}
\alias{is.named.list}
\alias{is.nonempty.string}
\alias{is.numeric.vector}
\alias{is.plain.list}
\alias{is.real.number}
\alias{is.vector.of.nonempty.strings}
\title{Various tests.}
\usage{
is.character.vector(x)

is.nonempty.string(x)

is.vector.of.nonempty.strings(x)

is.plain.list(x)

is.named.list(x)

is.integer.valued(x)

is.logical.vector(x)

is.empty.string(x)

\method{is.numeric}{vector}(x)

is.real.number(x)

is.inf(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
These are typically used in statements with \code{assert_that(...)}.
Each function returns either \code{TRUE} or \code{FALSE}; some functions
may return \code{NA}.

#' \itemize{
  \item{\code{is.character.vector}}: \code{TRUE} if and only if \code{x} is a
    character vector of positive length, otherwise \code{FALSE}.
  \item{\code{is.nonempty.string}}: \code{TRUE} if and only if \code{x} is a
    character vector of length 1 with a component that is nonempty and non-NA,
    otherwise \code{FALSE}.
  \item{\code{is.vector.of.nonempty.strings}}: \code{TRUE} if and only if
    \code{x} is a vector of nonempty strings none of which is an NA, otherwise
    \code{FALSE}.  Empty vectors yield \code{FALSE}.
  \item{\code{is.plain.list}}: \code{TRUE} if and only if \code{x} is a list
    but does not have the object class set. Otherwise \code{FALSE}.
  \item{\code{is.plain.list}}: \code{TRUE} if and only if \code{x} is a list
    but does not have the object class set. Otherwise \code{FALSE}.
  \item{\code{is.named.list}}: \code{TRUE} if and only if \code{x} is a plain
    list with a names attribute with non-empty strings for names, otherwise
    \code{FALSE}.
  \item{\code{is.integer.valued}}: \code{TRUE} if and only if \code{x} is
    either of type integer, or numeric and all components coercible to
    integer, otherwise \code{FALSE}.  NAs are allowed.
  \item{\code{is.logical.vector}}: \code{TRUE} if and only if \code{x} is a
    logical vector of positive length, otherwise \code{FALSE}.
  \item{\code{is.empty.string}}: A logical vector of the same length as
    \code{x}, #' with '\code{TRUE}' if and only if the string is empty,
    NA if the string is NA, otherwise \code{FALSE}. If \code{x} is not
    character, throws an error.  NA maps to NA.
  \item{\code{is.numeric.vector}}: \code{TRUE} if and only if \code{x} is a
    numeric vector of positive length, otherwise \code{FALSE}.
  \item{\code{is.real.number}}: \code{TRUE} if and only if \code{x} is a
    numeric scalar, not an NA, and not infinite, otherwise \code{FALSE}.
  \item{\code{is.inf}}: \code{TRUE} if and only if \code{x} is \code{Inf} or
    \code{-Inf}.
}
}
\description{
Various tests.
}

