% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_vectors.R
\name{is.treatment.assignment}
\alias{is.geo.group.number}
\alias{is.period.number}
\alias{is.treatment.assignment}
\title{Tests each component of the vector 'x' for valid values.}
\usage{
is.treatment.assignment(x)

is.geo.group.number(x)

is.period.number(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
A logical vector of the same length as 'x', with 'TRUE' if and only
if the string is a valid value, NA if the string is NA, otherwise FALSE. If
'x' is not character, throws an error.
}
\description{
Tests each component of the vector 'x' for valid values.
}
\note{
NA in a component returns NA. These are supposed to be tested
separately.

\itemize{
  \item{\code{is.treatment.assignment}} The valid treatment assignment
    symbols can be found in the constant \code{kTreatmentAssignment}.
    Note however that the treatment assignment symbols are not used
    in the current version of the package (included for possible future
    use).
  \item{\code{is.geo.group.number}} The value '0' has a special meaning:
   it indicates a geo that has been omitted from the experiment.
  \item{\code{is.period.number}} Negative period numbers are allowed.
}
}

