% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{date_utilities}
\alias{.GetLastDayOfMonth}
\alias{.GetLastDayOfWeek}
\alias{.GetWeekIndex}
\alias{.GetWeekNumbers}
\alias{.GetWeekdays}
\alias{date_utilities}
\title{Utilities to manipulate Date objects.}
\usage{
.GetWeekdays(x)

.GetWeekIndex(x)

.GetWeekNumbers(x)

.GetLastDayOfWeek(x, last.day = 7L)

.GetLastDayOfMonth(x)
}
\arguments{
\item{x}{a Date object (vector).}

\item{last.day}{(integer) number of the day to translate the date to. Monday
is 1, Sunday is 7.}
}
\value{
\itemize{
  \item{\code{.GetWeekdays}}: Indicator for weekday, given a
    date: an integer vector of the same length as \code{x}, indicating
    the number of the weekday, Monday=1, Tuesday=2, ..., Sunday=7.
  \item{\code{.GetWeekIndex}}: An indicator for an 'absolute' week number:
     an integer vector of the same length as \code{x}, indicating the week
     number, composed of the 4-digit year and the number of the week within
     the year.
  \item{\code{.GetWeekNumbers}}: Indicators for week numbers within the
     year: an integer vector of the same length as 'x', indicating the week
     number, 0 .. 53.
  \item{code{.GetLastDayOfWeek}}: Translates the given date(s) to the date
    of the last day of the corresponding week: a date object of the same
    length as 'x', with the corresponding dates shifted to the next (future)
    day given by 'last.day'.
  \item{\code{.GetLastDayOfMonth}}: Translates the given date(s) to the last
    day of the month of the corresponding month.  A date object of the same
    length as 'x', with the corresponding dates shifted to the last day of
    the month.
}
}
\description{
Utilities to manipulate Date objects.
}
\note{
As a convention, a week starts on Monday and ends on Sunday.
}

