% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatetimeseries.R
\name{AggregateTimeseries}
\alias{AggregateTimeseries}
\title{Aggregate the metrics of a time series object over specified time
intervals.}
\usage{
AggregateTimeseries(obj, freq = c("weekly", "monthly"), ...)
}
\arguments{
\item{obj}{an object.}

\item{freq}{(string) 'weekly' or 'monthly' aggregation.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of the same class as 'obj'.
}
\description{
Aggregate the metrics of a time series object over specified time
intervals.
}
\note{
A generic S3 method.
}
\seealso{
AggregateTimeseries.GeoTimeseries.
}

