% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gbrroasanalysisdata.R
\name{as.GBRROASAnalysisData}
\alias{as.GBRROASAnalysisData}
\alias{as.GBRROASAnalysisData.GeoExperimentData}
\title{Coerces an object to a GBRROASAnalysisData object.}
\usage{
as.GBRROASAnalysisData(obj, ...)

\method{as.GBRROASAnalysisData}{GeoExperimentData}(obj,
  response = character(0), cost = character(0), pretest.period = 0L,
  intervention.period = 1L, cooldown.period = NULL, control.group = 1L,
  treatment.group = 2L, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{response}{(string) name of the response variable column.}

\item{cost}{(string) name of the cost variable column.}

\item{pretest.period}{(vector of non-negative integers) number(s) of the
period(s) forming the pretest period.}

\item{intervention.period}{(vector of non-negative integers) number(s) of
the period(s) forming the intervention period. All must be larger
than the largest period in the pretest period.}

\item{cooldown.period}{(NULL or vector of non-negative integers) number(s)
of the period(s) forming the cooldown period. All must be larger than
the largest period in the intervention period.}

\item{control.group}{(NULL or a vector of positive integers) number(s) of
geo groups forming the control group.}

\item{treatment.group}{(NULL or a vector of positive integers) number(s) of
geo groups forming the control group.}
}
\value{
A GBRROASAnalysisData object.
}
\description{
Coerces an object to a GBRROASAnalysisData object.
}
\seealso{
\code{\link{DoGBRROASAnalysis}}.
}

