% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_geoexperimentdata.R
\name{as.GeoExperimentData}
\alias{as.GeoExperimentData}
\title{Coerces an object to a GeoExperimentData object.}
\usage{
as.GeoExperimentData(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed to methods.}
}
\value{
A GeoExperimentData object.
}
\description{
Coerces an object to a GeoExperimentData object.
}
\note{
A generic S3 method.
}

