% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assert.R
\name{CheckForAllTrue}
\alias{CheckForAllTrue}
\title{Check whether all(x) is true.}
\usage{
CheckForAllTrue(x, ...)
}
\arguments{
\item{x}{(logical) vector. A 'TRUE' denotes a success, and 'FALSE' and 'NA'
a failure. The names attribute must be set, with nonempty, non-NA,
non-duplicated names.}

\item{...}{one or more character vectors, passed on to \code{FormatText}.}
}
\value{
If all tests pass, returns TRUE invisibly. Otherwise, an
  'assertError' is thrown.
}
\description{
Check whether all(x) is true.
}

