% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{ComputeLinearModelWeights}
\alias{ComputeLinearModelWeights}
\title{Computes the weights to be used in the weighted linear model used to
estimate ROAS.}
\usage{
ComputeLinearModelWeights(response, epsilon = 0.001, power = 2)
}
\arguments{
\item{response}{a vector of the response in the pre period. Length equal to
the number of geos. Must be all nonnegative.}

\item{epsilon}{(number) a small positive increment to add to 'response' to
avoid 1 / 0.}

\item{power}{default power to which 'response' is raised to. Can be
overridden by setting the global option
'geoexperiments.gbr1.weight.power'.}
}
\value{
A vector of weights of the same length as 'response'.
}
\description{
Computes the weights to be used in the weighted linear model used to
estimate ROAS.
}
\note{
If a component of 'response' tends to infinity, the corresponding
weight tends to 0 (i.e., the corresponding data point is ignored).
}

