% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogbrroasanalysis.R
\name{DoGBRROASAnalysis}
\alias{DoGBRROASAnalysis}
\title{Performs a GBR ROAS Analysis.}
\usage{
DoGBRROASAnalysis(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A GBRROASAnalysisFit object.
}
\description{
Performs a GBR ROAS Analysis.
}
\note{
A generic S3 method. See also: DoGBRROASAnalysis.GBRROASAnalysisData,
DoGBRROASAnalysis.GeoExperimentData.
}

