% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doroaspreanalysis.R
\name{DoROASPreanalysis}
\alias{DoROASPreanalysis}
\title{Predicts the standard deviation (standard error) of the posterior of the
iROAS estimator.}
\usage{
DoROASPreanalysis(obj, ...)
}
\arguments{
\item{obj}{some object.}

\item{...}{arguments passed on to the models.}
}
\value{
A ROASPreanalysisFit object.
}
\description{
Predicts the standard deviation (standard error) of the posterior of the
iROAS estimator.
}

