% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotbranalysis.R
\name{DoTBRAnalysis}
\alias{DoTBRAnalysis}
\title{Performs a TBR Causal Effect Analysis.}
\usage{
DoTBRAnalysis(obj, model, ...)
}
\arguments{
\item{obj}{an object.}

\item{model}{(string) model to use.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A TBRAnalysisFit object.
}
\description{
Performs a TBR Causal Effect Analysis.
}
\note{
A generic S3 method.
}
\seealso{
DoTBRAnalysis.TBRAnalysisData, DoTBRAnalysis.GeoExperimentData.
}

