% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractexperimentperiods.R
\name{ExtractExperimentPeriods}
\alias{ExtractExperimentPeriods}
\title{Extracts an ExperimentPeriods object.}
\usage{
ExtractExperimentPeriods(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed on to methods.}
}
\value{
An ExperimentPeriods object.
}
\description{
Extracts an ExperimentPeriods object.
}
\note{
A generic S3 method.
}
\seealso{
ExtractExperimentPeriods.GeoTimeseries.
}

