% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractgeos.R
\name{ExtractGeos}
\alias{ExtractGeos}
\title{Extract a Geos object from an object.}
\usage{
ExtractGeos(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{other arguments passed to the methods.}
}
\value{
A 'Geos' object.
}
\description{
Extract a Geos object from an object.
}

