% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentassignment.R
\name{ExtractTreatmentAssignment}
\alias{ExtractTreatmentAssignment}
\title{Extracts a TreatmentAssignment object.}
\usage{
ExtractTreatmentAssignment(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{further arguments passed on to methods.}
}
\value{
A TreatmentAssignment object.
}
\description{
Extracts a TreatmentAssignment object.
}
\note{
A generic S3 method.
}
\seealso{
SetTreatmentAssignment<-.GeoExperimentData.
}

