% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoexperimentdata.R
\name{GeoExperimentData}
\alias{GeoExperimentData}
\title{Constructs a GeoExperimentData object.}
\usage{
GeoExperimentData(geo.timeseries, periods, geo.assignment, treat.assignment)
}
\arguments{
\item{geo.timeseries}{a GeoTimeseries object.}

\item{periods}{an \code{ExperimentPeriods} object, specifying the start and
end dates of each period in the experiment. Or, \code{NULL} if the dates are yet
unknown. Or, leave unspecified to extract this information from the column
'periods' of the \code{geo.timeseries}; in this case the column \emph{must}
exist in the data frame.}

\item{geo.assignment}{a \code{GeoAssignment} object, specifying the mapping
from a geo to a geo group. Or, \code{NULL} if the geo assignment is yet
unknown.  Or, leave unspecified to extract this information from the column
'geo.group' of \code{geo.timeseries}; in this case the column \emph{must}
exist in the data frame.}

\item{treat.assignment}{a \code{TreatmentAssignment} object, specifying the
mapping from (period, geo) to a treatment assignment condition (treatment
intervention type). Or, \code{NULL} if the treatment assignment is yet
unknown. Or, leave unspecified to extract this information from the column
\code{assignment} of \code{geo.timeseries}.  If the column \code{assignment}
does not exist, the resulting object will have the column \code{assignment}
filled with \code{NA}s.}
}
\value{
A GeoExperimentData object.
}
\description{
Constructs a GeoExperimentData object.
}
\details{
If any of the arguments \code{periods}, \code{geo.assignment}, or
\code{treat.assignment} is \code{NULL}, the corresponding column
(\code{period}, \code{geo.group}, \code{assignment}) in the resulting
\code{GeoExperimentData} object will have only \code{NA}s.\cr

If \code{geo.timeseries} has columns \code{period}, \code{geo.group}, or
\code{assignment}, they will be overwritten by the values specified by
\code{periods}, \code{geo.assignment}, and \code{treat.assignment}.\cr

The resulting object \emph{may} have undefined \code{periods},
\code{geo.group}, or \code{assignment}.  The missing parts may be filled in
by using the functions \code{SetExperimentPeriods}, \code{SetGeoAssignment},
and \code{SetTreatmentAssignment}.
}

