% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotimeseries.R
\name{GeoTimeseries}
\alias{GeoTimeseries}
\title{Constructs a GeoTimeseries object.}
\usage{
GeoTimeseries(x, metrics = character(0), date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{a _plain_ 'data.frame' or an object that has to be coercible to a
_plain_ 'data.frame' with columns 'date', 'geo'. 'date' must be a
Date object or a character vector or factor coercible to Date, and
'geo' must be a character vector, factor, or integer-valued. All
columns that start with a dot ('.') are removed. If 'geo' is an
integer-valued numeric or factor, it is coerced to character silently
without error. If 'date' is a character or factor, it is silently
coerced to Date without error. An error is output if the date
conversion fails.}

\item{metrics}{(character) column names that point to numeric columns. At
least one metric must be specified. All metrics have to be not all
NA.}

\item{date.format}{(string, optional) format of the column 'date' in the
form understood by as.Date(). Used only if the 'date' column is of
character type.}
}
\value{
A GeoTimeseries object.
}
\description{
Constructs a GeoTimeseries object.
}
\note{
'GeoTimeseries' is a 'data.frame', with the required columns,
\itemize{
\item\code{date} a 'Date' vector.
\item\code{geo} a 'character'-valued vector of Geo IDs.
\item metrics : one or more numeric metrics.} In addition, optionally
any other user-definable columns. (These are ignored by the specified
methods but may be convenient for the user). Three columns are added,
for convenience:
\itemize{
  \item\code{.weekday} : day number (1 = Monday, 2 = Tuesday, ..., 7 =
    Sunday)
  \item\code{.weeknum} : week number in the year from 0 to 53.
    \item\code{.weekindex} : absolute index of week, year + weeknum
    (e.g., '201542')
}
These should be convenient for generating totals and averages, using the
\code{aggregate} method. The columns 'date' and 'geo' form the primary
keys: it is guaranteed that no duplicate ('date', 'geo') pairs exist.
The object includes fields stored in the attribute 'info':
\itemize{
  \item\code{metrics}: names of the metric columns.
  \item\code{other}: names of the other user-supplied columns.
}
}

