% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeonames.R
\name{GetGeoNames}
\alias{GetGeoNames}
\title{Extracts the names of the geos from the object.}
\usage{
GetGeoNames(obj, groups = NULL)
}
\arguments{
\item{obj}{an object.}

\item{groups}{(NULL, or integer vector) id number(s) of the groups whose
geos to obtain, or NULL for all geos. NA is allowed.}
}
\value{
A character vector of unique geo identifiers.
}
\description{
Extracts the names of the geos from the object.
}

