% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostrata.R
\name{IsFixedRandomization}
\alias{IsFixedRandomization}
\title{Returns TRUE iff randomizing the geostrata can lead to multiple outcomes.}
\usage{
IsFixedRandomization(geostrata)
}
\arguments{
\item{geostrata}{a GeoStrata object.}
}
\value{
TRUE if randomizing the geostrata can only lead to a single
  GeoAssignment, FALSE otherwise.
}
\description{
Returns TRUE iff randomizing the geostrata can lead to multiple outcomes.
}

