% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isinperiod.R
\name{IsInPeriod}
\alias{IsInPeriod}
\title{[internal] Returns a logical vector indicating which of the rows are in
the specified period(s).}
\usage{
IsInPeriod(obj, periods)
}
\arguments{
\item{obj}{some object.}

\item{periods}{names of the periods.}
}
\value{
A logical vector of length \code{nrow(obj)}, TRUE for a row that is
  within the given period(s), FALSE otherwise. No NAs.
}
\description{
[internal] Returns a logical vector indicating which of the rows are in
the specified period(s).
}

