% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{Randomize}
\alias{Randomize}
\title{Randomize geos to groups.}
\usage{
Randomize(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{arguments passed to other methods.}
}
\value{
A GeoAssignment object.
}
\description{
Randomize geos to groups.
}

