% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulategeoxdataset.R
\name{SimulateGeoExperimentData}
\alias{SimulateGeoExperimentData}
\title{Returns a geo experiment data set.}
\usage{
SimulateGeoExperimentData(obj, ...)
}
\arguments{
\item{obj}{a GeoExperimentPreanalysisData object.}

\item{...}{arguments passed on to the methods.}
}
\value{
A GeoExperimentData object.
}
\description{
Returns a geo experiment data set.
}

