% Copyright (C) 2017 Google, Inc.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentassignment.R
\name{TreatmentAssignment}
\alias{DefaultTreatmentAssignment}
\alias{TreatmentAssignment}
\title{Constructs a TreatmentAssignment object.}
\usage{
TreatmentAssignment(x)

DefaultTreatmentAssignment()
}
\arguments{
\item{x}{(data frame) a mapping from (period, group) to treatment assignment
condition. The data frame must have the columns 'period'
(integer-valued), 'geo.group' (integer-valued, positive), and
'assignment' (integer-valued, one of 0, 1, -1), Each row specifies
the mapping of one (period, group) pair. No missing values are
allowed.}
}
\value{
An object of class 'TreatmentAssignment'.
}
\description{
Constructs a TreatmentAssignment object.
}
\note{
Not used in the analysis methods, only for
\code{\link{SetSpendChange<-}}.

\code{DefaultTreatmentAssignment} generates the default treatment
assignment condition, which assigns a change ('some intervention') to Period
1 of group 2.
}
\seealso{
\code{\link{TreatmentAssignment}}, \code{\link{SetSpendChange<-}}.
}

